+++ paper/paper.md
@@ -1,0 +1,131 @@
+---
+title: TiLiA – TimeLineAnnotator
+tags:
+  - music
+  - scores
+  - film
+  - visualisation
+  - data science
+authors:
+  - name: Felipe D. Martins
+    orcid: https://orcid.org/0000-0002-9965-042X
+    equal-contrib: true
+    affiliation: 1
+  - name: Anne R. E. Foo
+    orcid: https://orcid.org/0009-0003-9839-473X
+    equal-contrib: true
+    affiliation: 2
+  - name: Mark R. H. Gotham
+    orcid: https://orcid.org/0000-0003-0722-3074
+    corresponding: true # (This is how to denote the corresponding author)
+    equal-contrib: true
+    affiliation: 3
+affiliations:
+ - name: Universidade Federal do Rio de Janeiro, Brazil
+   index: 1
+ - name: Independent Researcher, United Kingdom
+   index: 2
+ - name: King's College London, United Kingdom
+   index: 3
+date: 2025-02
+bibliography: paper.bib
+
+---
+
+
+# Summary
+
+'TiLiA' (**Ti**me**Li**ne **A**nnotator) is an open-source, cross-platform network of tools and resources designed to facilitate real-time creating, displaying and interacting with timeline-style annotations of audio, video and musical scores. Developed in Python with the PyQt UI framework, TiLiA is available primarily as a desktop Graphical User Interface (GUI),supported by a Command Line Interface (CLI) as well as a free and open online platform where users may upload, view and query existing analyses.
+
+# Statement of need
+
+Visual analyses of music have proven to be an enduringly popular, effective, and versatile tool for conveying a variety of musical ideas. [^0] As the famous adage goes: 'a picture is worth a thousand words'. These analytical images are primarily connected to the *analysis* rather than the *music*, thus making them readily useful not only for the already visual sources of notated music, but also for the much broader range of musical styles and repertoires that are transmitted without the use of notation.
+
+[^0]: See @isaacson:2023 for a survey of this long history.
+
+Whatever potential a *static* image has, an *interactive* one offers considerably more. Formal analysis is one of the clearest beneficiaries of the visual summary of music. Form can be hard to parse (especially in real-time) but is often easy to understand in relatively simple, at-a-glance summaries.
+
+Save for a few exceptions [@gotham:2019], musical scholarship has not yet made the leap into this technological space with any degree of decisiveness or consistency.
+Even most visually minded music analyses are consigned to static images embedded in print (or in e-copies at best) and not subject to the expectations for _fair_,
+open-source data sharing that are now *de rigueur* in other fields.
+Still among the more computationally inclined communities, there has been little progress made on the coordination of standards and corpora for formal analysis.
+
+In today's digital age, we have come to expect certain basic interactive
+features across visual interfaces, such as **resizing**
+content to fit the window, keyboard shortcuts, undo/redo functionality, and so on. Musical contexts extend this with additional
+expectations, such as **real-time synchronisation** (a.k.a., audio or
+score following). Any tool that aims to win over users effectively must present analytical ideas naturally and intuitively, with minimal friction in the user interface.
+
+While one _can_ appropriate tools (which primary purpose is editing scores or audio) for creating visual analyses, [^1] [^2] the current range of tools dedicated to the task typically offer limited functionality and their uptake by the wider community haws been minimal. [^3]
+
+[^1]: The music notation software _Sibelius_, for instance, has offered a timeline since c.2019 with bar numbers and rudimentary checkpoints for changes to time and key signatures, presumably because they are possible indicators of section breaks. Peebles [@peebles:2013] discusses how to use an audio editor (_Audacity_) in a similar way to the tools discussed here.
+
[…]
