+++ paper/paper.bib
@@ -1,0 +1,106 @@
+@book{isaacson:2023,
+	title = {Visualizing {Music}},
+	url = {https://www.barnesandnoble.com/w/visualizing-music-eric-isaacson/1140906228?ean=9780253064738},
+	language = {en},
+	urldate = {2023-05-08},
+	publisher = {Indiana University Press},
+	author = {Isaacson, Eric},
+	year = {2023},
+	file = {Snapshot:/Users/k2476966/Zotero/storage/BRVMMEU5/1140906228\;jsessionid=39750F12CB32F82156C041A0CF91C6A7.html:text/html},
+}
+
+@inproceedings{gotham:2019,
+	title = {Taking {Form}: {A} {Representation} {Standard}, {Conversion} {Code}, and {Example} {Corpora} for {Recording}, {Visualizing}, and {Studying} {Analyses} of {Musical} {Form}},
+	copyright = {All rights reserved},
+	url = {http://archives.ismir.net/ismir2019/paper/000084.pdf},
+	booktitle = {Proceedings of the 20th {International} {Society} for {Music} {Information} {Retrieval} {Conference}, {ISMIR} 2019, {Delft}, {The} {Netherlands}, {November} 4-8, 2019},
+	author = {Gotham, Mark R. H. and Ireland, Matthew},
+	editor = {Flexer, Arthur and Peeters, Geoffroy and Urbano, Julián and Volk, Anja},
+	year = {2019},
+	keywords = {Corpus/Data(sets), Analysis, Digital / Computational, Form},
+	pages = {693--699},
+}
+
+@article{peebles:2013,
+	title = {Using {Audacity} to {Participate} in {Active} {Musical} {Listening}},
+	url = {http://flipcamp.org/engagingstudents/peebles.html},
+	urldate = {2023-05-08},
+	journal = {Engaging Students: Essays in Music Pedagogy},
+	author = {Peebles, Crystal},
+	year = {2013},
+	file = {Engaging Students\: Essays in Music Pedagogy:/Users/k2476966/Zotero/storage/9KZG2MIR/peebles.html:text/html},
+}
+
+@misc{rowenstrunk:2014,
+	title        = {Markieren, {Codieren}, {Annotieren} – {Arbeiten} an und mit digitalen {Musikeditionen} mit {Edirom} und {MEI}},
+	author       = {Röwenstrunk, Daniel},
+	year         = 2014,
+	month        = jun,
+	address      = {Bern, Schweiz},
+	keywords     = {Edirom, MEI, ViFE}
+}
+
+@inproceedings{weigl:2023,
+	title        = {Listen Here! A Web-native digital musicology environment for machine-assisted close listening},
+	author       = {Weigl, David M. and VanderHart, Chanda and Rammler, Delilah and Pescoller, Matth\"{a}us and Goebl, Werner},
+	year         = 2023,
+	booktitle    = {Proceedings of the 10th International Conference on Digital Libraries for Musicology},
+	location     = {Milan, Italy},
+	publisher    = {Association for Computing Machinery},
+	address      = {New York, NY, USA},
+	series       = {DLfM '23},
+	pages        = {109–118},
+	doi          = {10.1145/3625135.3625144},
+	isbn         = 9798400708336,
+	url          = {https://doi.org/10.1145/3625135.3625144},
+	numpages     = 10,
+	keywords     = {performance recordings, music encoding, music annotation, linked data, digital musicology, close listening}
+}
+
+@article{goebl:2024,
+	title        = {mei-friend: An Interactive Web-based Editor for Digital Music Encodings},
+	author       = {Goebl,Werner  and Weigl, David M.},
+	year         = 2024,
[…]
