### 2025-03-12 d11c529
fix: using iterable as source of return values for patch.object
