### 2025-03-12 134807a
test: add patch function for ask_for_string dialog
