### 2025-03-12 0c46a00
refactor: remove duplication
