### 2025-03-12 1a704f0
fix: do not export certain attributes
