### 2025-03-12 69d2258
fix: do not export certain attributes
