### 2025-03-12 4d50cd6
fix: do not export certain attributes
