+++ tilia/timelines/score/timeline.py
@@ -47,6 +47,7 @@ class ScoreTimeline(Timeline):
         "svg_data",
         "viewer_beat_x",
     ]
+    NOT_EXPORTABLE_ATTRS = ["svg_data", "viewer_beat_x"]
     COMPONENT_MANAGER_CLASS = ScoreTLComponentManager
 
     def __init__(
