### 2025-03-12 c891d6c
test: timeline attributes are exported correctly
