### 2025-03-12 b5d4fce
test: timeline attributes are exported correctly
