+++ tests/test_export.py
@@ -2,9 +2,12 @@ import json
 
 from tests.conftest import parametrize_component
 from tests.constants import EXAMPLE_MEDIA_PATH
-from tests.mock import Serve
+from tests.mock import Serve, patch_ask_for_string_dialog
 from tilia.requests import post, Post, Get
+from tilia.settings import settings
 from tilia.timelines.base.timeline import TimelineFlag
+from tilia.timelines.harmony.timeline import HarmonyTimeline
+from tilia.timelines.timeline_kinds import TimelineKind
 from tilia.ui.actions import TiliaAction
 
 
@@ -16,6 +19,50 @@ def _trigger_export_action(user_actions, path):
         data = json.load(f)
 
     return data
+
+
+def test_timelines_attributes_are_exported(
+    tilia, marker_tlui, user_actions, tmp_path, use_test_settings
+):
+    # Marker timeline is chosen as an example. Ideally this should be parametrized for all timelines kinds.
+    tl_name = "my name"
+    with patch_ask_for_string_dialog(True, tl_name):
+        user_actions.trigger(TiliaAction.TIMELINE_NAME_SET)
+
+    tmp_file = tmp_path / "test.json"
+
+    data = _trigger_export_action(user_actions, tmp_file)
+
+    timeline_data = data["timelines"][0]
+    assert "hash" not in timeline_data
+    assert timeline_data["ordinal"] == 1
+    assert timeline_data["name"] == tl_name
+    assert timeline_data["is_visible"] == True
+    assert timeline_data["height"] == settings.get("marker_timeline", "default_height")
+    assert timeline_data["kind"] == TimelineKind.MARKER_TIMELINE.name
+    assert timeline_data["components"] == []
+
+
+def test_timeline_not_exportable_attributes_are_not_exported(
+    tilia, harmony_tlui, user_actions, tmp_path
+):
+    # Harmony timeline is chosen as an example. Ideally this should be parametrized for all timelines kinds that have not exportable attributes.
+    tmp_file = tmp_path / "test.json"
+    data = _trigger_export_action(user_actions, tmp_file)
+
+    timeline_data = data["timelines"][0]
+    for attr in HarmonyTimeline.NOT_EXPORTABLE_ATTRS:
+        assert attr not in timeline_data
+
+
+def test_not_exportble_timelines_are_not_exported(
+    tilia, marker_tlui, audiowave_tlui, user_actions, tmp_path
+):
+    # Audiowave timeline is chosen as an example. Ideally this should be parametrized for all non-exportable timeline types.
+    tmp_file = tmp_path / "test.json"
+    data = _trigger_export_action(user_actions, tmp_file)
+
+    assert len(data["timelines"]) == 1
 
 
 def test_export_timelines(
