### 2025-03-12 cd42dc9
[SQUASH] fix: outdated tests
