### 2025-03-12 9591aec
chore: add flake8 pre-commit hook
