+++ .pre-commit-config.yaml
@@ -13,3 +13,8 @@ repos:
         args: ['--maxkb=1000']
       - id: debug-statements
       - id: check-yaml
+
+-   repo: https://github.com/pycqa/flake8
+    rev: 7.1.0
+    hooks:
+    -   id: flake8
