### 2025-03-12 12e046b
build: continue other tests on failure
