+++ .github/workflows/run-tests.yml
@@ -12,6 +12,7 @@ jobs:
   build:
     runs-on: ${{ matrix.os }}
     strategy:
+      fail-fast: false
       matrix:
         # TODO: test for 3.13
         python-version: ['3.10', '3.11', '3.12']
@@ -59,13 +60,14 @@ jobs:
       run: |
         pip freeze
 
-    - name: Test with pytest
-      run: |
-          coverage run -m pytest -v -s
-
     - name: Lint with flake8
       run: |
         flake8
+      continue-on-error: true
+
+    - name: Test with pytest
+      run: |
+        coverage run -m pytest -v -s
 
     - name: Generate Coverage Report
       run: |
