### 2025-03-13 d2c7522
fix: merged time signatures
Time signatures were overlapping due to not setting the right staff index
