### 2025-03-13 f238c35
fix: etree FutureWarning
