+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = tilia
-version = 0.5.9
+version = 0.5.10
 description = A GUI-based software for creating and visualizing annotations over audio and video files.
 author = Felipe Defensor
 
