### 2025-03-13 5c28765
chore: improve Clef.__str__
