### 2025-03-13 21b54aa
chore: more informative TimelineUIElement.__repr__
