### 2025-03-13 99f6f6f
build: don't run tests on documentation updates
