+++ .github/workflows/run-tests.yml
@@ -5,8 +5,17 @@ on:
     branches:
       - main
       - dev
+    paths-ignore:
+      - 'docs/**'
+      - 'paper/**'
+      - '**.md'
+
   pull_request:
     types: [opened, reopened, synchronize]
+    paths-ignore:
+      - 'docs/**'
+      - 'paper/**'
+      - '**.md'
 
 jobs:
   build:
