### 2025-03-13 51fdf89
build: add permissions
