### 2025-03-14 d049a5d
test: fix failure on 3.10
For some reason, Python 3.10 only takes iterables as a `side_effect` argument for `unittest.patch.object`. I couldn't find anything on the for 3.11 about changing or fixing this behavior.
