### 2025-03-14 ee4674f
fix: incorrect parsing of clef
