### 2025-03-17 0d09748
perf: defer BarLineUI body creation
