+++ setup.cfg
@@ -1,5 +1,5 @@
 [metadata]
-name = tilia
+name = TiLiA
 version = 0.5.10
 description = A GUI-based software for creating and visualizing annotations over audio and video files.
 author = Felipe Defensor
