### 2025-03-18 5c688de
feat: make license viewable from in-app
