### 2025-03-18 0d4b0db
chore: licensing
include licensing notice
qtui: exclude licensing instructions in display
