### 2025-03-18 2a46cb0
chore: licensing
include licensing notice
qtui: exclude licensing instructions in display
