+++ tilia/ui/windows/about.py
@@ -8,7 +8,7 @@ from PyQt6.QtWidgets import (
     QScrollArea,
     QVBoxLayout,
 )
-from re import sub
+from re import split, sub
 
 import tilia.constants
 
@@ -37,7 +37,7 @@ class About(QDialog):
         gh_label.setTextFormat(Qt.TextFormat.RichText)
         gh_label.setAlignment(Qt.AlignmentFlag.AlignCenter)
         license_label = QLabel(
-            f'<a href="#license">{tilia.constants.APP_NAME} Copyright © 2025 {tilia.constants.AUTHOR}</a>'
+            f'<a href="#license">{tilia.constants.APP_NAME} Copyright © {tilia.constants.YEAR} {tilia.constants.AUTHOR}</a>'
         )
         license_label.setAlignment(Qt.AlignmentFlag.AlignCenter)
         license_label.linkActivated.connect(self.open_link)
@@ -69,17 +69,24 @@ class License(QDialog):
         layout = QVBoxLayout()
         self.setLayout(layout)
 
+        notice = QLabel(tilia.constants.NOTICE)
+        notice.setWordWrap(True)
+
         with open(
             Path(__file__).parent.parent.parent.parent / "LICENSE", encoding="utf-8"
         ) as license_file:
-            text = f"<pre style='white-space: pre-wrap;'>{license_file.read()}</pre>"
-        text = sub(
+            license_text = split(
+                "How to Apply These Terms to Your New Programs", license_file.read()
+            )[0]
+
+        formatted_text = f'<pre style="white-space: pre-wrap;">{license_text}</pre>'
+        text_with_links = sub(
             "<https:([^>]+)>",
             lambda y: f'<a href="{y[0][1:-1]}">{y[0][1:-1]}</a>',
-            text,
+            formatted_text,
         )
 
-        license_text = QLabel(text)
+        license_text = QLabel(text_with_links)
         license_text.setTextFormat(Qt.TextFormat.RichText)
         license_text.setOpenExternalLinks(True)
         license_text.setContentsMargins(5, 5, 5, 5)
@@ -94,5 +101,6 @@ class License(QDialog):
         license_scroll.setFrameShadow(QFrame.Shadow.Sunken)
         license_scroll.setFrameShape(QFrame.Shape.Panel)
 
+        layout.addWidget(notice)
         layout.addWidget(license_scroll)
         layout.setSpacing(5)
