+++ setup.cfg
@@ -1,8 +1,8 @@
 [metadata]
 name = TiLiA
 version = 0.5.10
-description = A GUI-based software for creating and visualizing annotations over audio and video files.
-author = Felipe Defensor
+description = A GUI for creating and visualizing annotations over audio and video files.
+author = Felipe Defensor et al.
 
 [options]
 packages = tilia
