### 2025-03-20 #7893 @azfoo
@arfon: We have cleared up the issue with the paper's branch and the license and are ready to continue with the process.
(Felipe is away for a while so I (2nd author) am taking over temporarily.)

Regarding reviewers,  Maik Köster (m-k94) is fairly familiar with the app from the perspective of a user -- we have no prior collaboration but are in the midst of planning a paper on dataset alignment, tangentially related to the work of this software.
Other potential reviewers are Stefan Balke (stefan-balke), Magdalena Fuentes (magdalenafuentes) and Brian McFee (bmcfee).
