### 2025-03-22 b6c3fd2
fix: make sure function returns a value
