### 2025-03-22 67bd954
refactor: return earlier
