### 2025-03-22 c2a0bc5
refactor: reuse existing function
