### 2025-03-22 fdb76eb
fix: allow empty staves
