### 2025-03-22 a637359
fix: allow empty staves
