### 2025-03-22 6222da2
fix: backslashes on Windows creating duplicate entries
