### 2025-03-24 0018c2b
fix: validate staff numbers
ensure no duplicate or missing staff import
