### 2025-03-24 c1c516d
test: correct clef, key and time signature position
