### 2025-03-24 6706f08
test: correct clef, key and time signature position
