### 2025-03-24 01be330
test: pushing attributes in multi-stave score
