### 2025-03-26 0811755
fix: incorrect staff numbers
- missing staff numbers cause a crash and duplicates don't print correctly
- delete timeline if incorrect staff number to prevent crash
