### 2025-03-26 b3129ab
fix: correct margins for empty staves
empty staves with symbols need margins on the top so symbols do not collide with the staff
