### 2025-03-26 39e01d9
test: missing/duplicate staff components deletes timeline
