### 2025-03-26 bccf4e5
test: staff margins
- staff with no notes does not collide with symbols

- no extra margin is added for staff without symbols
