+++ tests/ui/timelines/score/test_score_timeline_ui.py
@@ -2,7 +2,7 @@ import pytest
 import json
 
 from tests.constants import EXAMPLE_MULTISTAFF_MUSICXML_PATH
-from tests.mock import Serve, patch_file_dialog
+from tests.mock import Serve, patch_file_dialog, patch_yes_or_no_dialog
 from tests.utils import reloadable, get_blank_file_data
 from tilia.errors import SCORE_STAFF_ID_ERROR
 from tilia.parsers.score.musicxml import notes_from_musicXML
@@ -204,3 +204,89 @@ def test_duplicate_staff_deletes_timeline(
 
     tilia_errors.assert_in_error_title(SCORE_STAFF_ID_ERROR.title)
     assert tls.get_timeline_by_attr("KIND", TimelineKind.SCORE_TIMELINE) is None
+
+
+def test_symbol_staff_collision(qtui, tmp_path, user_actions):
+    file_data_with_symbols = get_blank_file_data()
+    file_data_with_symbols["timelines"] = {
+        0: {
+            "kind": "SCORE_TIMELINE",
+            "height": 1,
+            "is_visible": True,
+            "name": "",
+            "ordinal": 1,
+            "svg_data": "",
+            "viewer_beat_x": {},
+            "hash": "",
+            "components": {
+                1: {"line_count": 5, "index": 0, "kind": "STAFF", "hash": ""},
+                2: {
+                    "staff_index": 0,
+                    "time": 0,
+                    "line_number": -1,
+                    "step": 4,
+                    "octave": 4,
+                    "icon": "clef-treble.svg",
+                    "kind": "CLEF",
+                    "hash": "",
+                },
+            },
+            "components_hash": "",
+        }
+    }
+
+    tmp_file_with_symbols = tmp_path / "test_with_sym.tla"
+    tmp_file_with_symbols.write_text(
+        json.dumps(file_data_with_symbols), encoding="utf-8"
+    )
+
+    with (patch_file_dialog(True, [tmp_file_with_symbols])):
+        user_actions.trigger(TiliaAction.FILE_OPEN)
+
+    score = get(Get.TIMELINE_UI_BY_ATTR, "TIMELINE_KIND", TimelineKind.SCORE_TIMELINE)
+    clef = score.timeline.get_component_by_attr("KIND", ComponentKind.CLEF)
+    staff = score.timeline.get_component_by_attr("KIND", ComponentKind.STAFF)
+
+    staff_top_y_with_symbols = (
+        score.get_element(staff.id).staff_lines.lines[0].line().y1()
+    )
+
+    assert score.get_element(clef.id).body.y() != staff_top_y_with_symbols
+
+    file_data_sans_symbols = get_blank_file_data()
+    file_data_sans_symbols["timelines"] = {
+        0: {
+            "kind": "SCORE_TIMELINE",
+            "height": 1,
+            "is_visible": True,
+            "name": "",
+            "ordinal": 1,
+            "svg_data": "",
+            "viewer_beat_x": {},
+            "hash": "",
+            "components": {
+                1: {"line_count": 5, "index": 0, "kind": "STAFF", "hash": ""},
+            },
+            "components_hash": "",
+        }
+    }
+
+    tmp_file_sans_symbols = tmp_path / "test_sans_sym.tla"
+    tmp_file_sans_symbols.write_text(
+        json.dumps(file_data_sans_symbols), encoding="utf-8"
+    )
+
+    with (
+        patch_file_dialog(True, [tmp_file_sans_symbols]),
+        patch_yes_or_no_dialog(False),  # do not save changes
+    ):
[…]
