### 2025-03-27 b633afc
test: use lower-level patch
