### 2025-03-27 3ed8526
test: use lower-level patch
