### 2025-03-28 9b4f5ad
feat: add email
required as part of licensing
