### 2025-03-28 71a4c63
feat: add email
required as part of licensing
