### 2025-03-28 97c04d6
fix: missing method
Staff does not inherit from PointLike or SegmentLike which has implementations of this method.
