+++ tilia/ui/dialogs/file.py
@@ -20,10 +20,13 @@ def ask_should_save_changes():
 
 
 def _get_return_from_file_dialog(dialog: QFileDialog):
-    return dialog.exec(), dialog.selectedFiles()[0] if dialog.selectedFiles() else None
+    return (
+        bool(dialog.exec()),
+        dialog.selectedFiles()[0] if dialog.selectedFiles() else None,
+    )
 
 
-def ask_for_tilia_file_to_open():
+def ask_for_tilia_file_to_open() -> tuple[bool, str | None]:
     dialog = QFileDialog()
     dialog.setWindowTitle(f"Open {APP_NAME} files")
     dialog.setFileMode(QFileDialog.FileMode.ExistingFile)
@@ -32,7 +35,9 @@ def ask_for_tilia_file_to_open():
     return _get_return_from_file_dialog(dialog)
 
 
-def ask_for_file_to_open(title: str, name_filters: str | Sequence[str]):
+def ask_for_file_to_open(
+    title: str, name_filters: str | Sequence[str]
+) -> tuple[bool, str | None]:
     dialog = QFileDialog()
     dialog.setWindowTitle(title)
     dialog.setFilter(QDir.Filter.Files)
@@ -45,7 +50,7 @@ def ask_for_file_to_open(title: str, name_filters: str | Sequence[str]):
 
 def ask_for_path_to_save(
     title: str, filter: str, initial_filename: str
-) -> tuple[bool, str]:
+) -> tuple[bool, str | None]:
     dialog = QFileDialog()
     dialog.setWindowTitle(title)
     dialog.setFileMode(QFileDialog.FileMode.AnyFile)
@@ -61,11 +66,13 @@ def ask_for_path_to_save(
     return _get_return_from_file_dialog(dialog)
 
 
-def ask_for_path_to_save_tilia_file(initial_filename: str) -> tuple[bool, str]:
+def ask_for_path_to_save_tilia_file(initial_filename: str) -> tuple[bool, str | None]:
     return ask_for_path_to_save("Save as", APP_FILE_FILTER, initial_filename)
 
 
-def ask_for_path_to_save_ogg_file(title: str, initial_name: str) -> tuple[bool, str]:
+def ask_for_path_to_save_ogg_file(
+    title: str, initial_name: str
+) -> tuple[bool, str | None]:
     return ask_for_path_to_save(title, "OGG files (*.ogg)", initial_name)
 
 
@@ -73,7 +80,7 @@ def ask_for_path_to_export(initial_name: str) -> tuple[bool, str]:
     return ask_for_path_to_save("Export", "JSON files (*.json)", initial_name)
 
 
-def ask_for_pdf_file():
+def ask_for_pdf_file() -> tuple[bool, str | None]:
     dialog = QFileDialog()
     dialog.setWindowTitle("Choose PDF")
     dialog.setFilter(QtCore.QDir.Filter.Files)
@@ -94,7 +101,7 @@ def ask_retry_pdf_file():
     )
 
 
-def ask_for_media_file():
+def ask_for_media_file() -> tuple[bool, str | None]:
     def get_filetypes_str(formats: list):
         filetypes = ""
         for frmt in formats:
@@ -130,5 +137,7 @@ def ask_for_media_file():
     return _get_return_from_file_dialog(dialog)
 
 
-def ask_add_timeline_without_media() -> tuple[bool, AddTimelineWithoutMedia.Result]:
+def ask_add_timeline_without_media() -> tuple[
+    bool, AddTimelineWithoutMedia.Result | None
+]:
     return AddTimelineWithoutMedia.select()
