### 2025-03-29 c845925
fix: mystery black bar at the bottom of image exports
- additional 20 margin was causing miscalculation. removing the margin does not change any other functionality.
- takes into account that margins do not change on zoom
