### 2025-03-29 c31f4fa
refactor: make default image export jpeg
