### 2025-03-29 38ad18c
fix: make dialog properly modal
