### 2025-03-29 64fb100
fix: declutter export ui
- removes original dimensions label
- makes height label static, since height does not change on zooming
