### 2025-03-29 259a446
feat: alt-click to select without seeking
