### 2025-03-29 56307ec
feat: alt-click to select without seeking
