### 2025-03-29 50fa776
refactor: move test into class
