+++ tests/test_export.py
@@ -11,136 +11,131 @@ from tilia.timelines.timeline_kinds import TimelineKind
 from tilia.ui.actions import TiliaAction
 
 
-def _trigger_export_action(user_actions, path):
-    with Serve(Get.FROM_USER_EXPORT_PATH, (True, path)):
-        user_actions.trigger(TiliaAction.FILE_EXPORT_JSON)
+class TestExportJSON:
+    def _trigger_export_action(self, user_actions, path):
+        with Serve(Get.FROM_USER_EXPORT_PATH, (True, path)):
+            user_actions.trigger(TiliaAction.FILE_EXPORT_JSON)
 
-    with open(path, encoding="utf-8") as f:
-        data = json.load(f)
+        with open(path, encoding="utf-8") as f:
+            data = json.load(f)
 
-    return data
+        return data
 
+    def test_timelines_attributes_are_exported(
+        self, tilia, marker_tlui, user_actions, tmp_path, use_test_settings
+    ):
+        # Marker timeline is chosen as an example. Ideally this should be parametrized for all timelines kinds.
+        tl_name = "my name"
+        with patch_ask_for_string_dialog(True, tl_name):
+            user_actions.trigger(TiliaAction.TIMELINE_NAME_SET)
 
-def test_timelines_attributes_are_exported(
-    tilia, marker_tlui, user_actions, tmp_path, use_test_settings
-):
-    # Marker timeline is chosen as an example. Ideally this should be parametrized for all timelines kinds.
-    tl_name = "my name"
-    with patch_ask_for_string_dialog(True, tl_name):
-        user_actions.trigger(TiliaAction.TIMELINE_NAME_SET)
+        tmp_file = tmp_path / "test.json"
 
-    tmp_file = tmp_path / "test.json"
+        data = self._trigger_export_action(user_actions, tmp_file)
 
-    data = _trigger_export_action(user_actions, tmp_file)
+        timeline_data = data["timelines"][0]
+        assert "hash" not in timeline_data
+        assert timeline_data["ordinal"] == 1
+        assert timeline_data["name"] == tl_name
+        assert timeline_data["is_visible"] is True
+        assert timeline_data["height"] == settings.get(
+            "marker_timeline", "default_height"
+        )
+        assert timeline_data["kind"] == TimelineKind.MARKER_TIMELINE.name
+        assert timeline_data["components"] == []
 
-    timeline_data = data["timelines"][0]
-    assert "hash" not in timeline_data
-    assert timeline_data["ordinal"] == 1
-    assert timeline_data["name"] == tl_name
-    assert timeline_data["is_visible"] is True
-    assert timeline_data["height"] == settings.get("marker_timeline", "default_height")
-    assert timeline_data["kind"] == TimelineKind.MARKER_TIMELINE.name
-    assert timeline_data["components"] == []
+    def test_timeline_not_exportable_attributes_are_not_exported(
+        self, tilia, harmony_tlui, user_actions, tmp_path
+    ):
+        # Harmony timeline is chosen as an example. Ideally this should be parametrized for all timelines kinds that have not exportable attributes.
+        tmp_file = tmp_path / "test.json"
+        data = self._trigger_export_action(user_actions, tmp_file)
 
+        timeline_data = data["timelines"][0]
+        for attr in HarmonyTimeline.NOT_EXPORTABLE_ATTRS:
+            assert attr not in timeline_data
 
-def test_timeline_not_exportable_attributes_are_not_exported(
-    tilia, harmony_tlui, user_actions, tmp_path
-):
[…]
