### 2025-03-29 19125ef
SQUASH
I forgot that we don't change heights when zooming, so the height field is pointless.
Leaving the height field visible for now - that might be useful for people who want to export to a fixed dimension.
