### 2025-03-31 4618a2a
fix: Exception never raised
