### 2025-03-31 cf35b7c
fix: Exception never raised
