### 2025-03-31 3394e78
feat: automatically update metadata title field on save if not set
