### 2025-03-31 c633565
feat: automatically update metadata title field on save if not set
