### 2025-04-04 124aafe
fix: typo
was causing musicxml files to not show up on linux.

closes #330
