### 2025-04-04 32eb6b2
test: generalize click_timeline_ui for multiple modifiers
