### 2025-04-04 35ea37b
fix: cancel add harmony or mode
