### 2025-04-04 416e56c
perf: cache beat metric positions
This improves by ~10x the speed of adding a beat at the start of a beat timeline with 300 beats.
