### 2025-04-05 81e6a26
chore: return value, success from TimelineComponentManager.set_component_data
This value can be used by subclasses to determine if further computing is necessary
