### 2025-04-05 e24bbff
perf: pause metric fraction dict update when updating is_first_in_measure for all beats
