### 2025-04-05 f3a8ccd
perf: pause metric fraction dict update when updating is_first_in_measure for all beats
