+++ tilia/ui/timelines/beat/dialogs.py
@@ -1,4 +1,5 @@
 from tilia.ui.windows.beat_pattern import AskBeatPattern
+from tilia.ui.windows.fill_beat_timeline import FillBeatTimeline, BeatTimeline
 import tilia.errors
 
 
@@ -18,3 +19,9 @@ def ask_for_beat_pattern():
         return ask_for_beat_pattern()
     else:
         return True, list(map(int, result))
+
+
+def ask_beat_timeline_fill_method() -> (
+    tuple[bool, None | tuple[BeatTimeline, BeatTimeline.FillMethod, float]]
+):
+    return FillBeatTimeline.select()
