### 2025-04-08 20c086b
refactor: some input dialogs
- function name `result` overrides parent function in an incompatible manner
- move beat dialog into related file
