### 2025-04-08 31f197e
refactor: some input dialogs
- function name `result` overrides parent function in an incompatible manner
- move beat dialog into related file
