### 2025-04-09 87df0b3
test: radical parametrisation
currently failing because some clicks are not actually finding the item to click on.
