+++ tests/ui/timelines/test_timeline_ui.py
@@ -8,132 +8,130 @@ from tilia.timelines.timeline_kinds import TimelineKind
 from tilia.ui.actions import TiliaAction
 
 
-class TestAltControlSelect:
-    # ideally we would test this on every kind of element
-    # but I didn't find an abstract way to click on all of them
+def get_time_shifted_args(args: dict[str, int], d_time) -> dict[str, int]:
+    if "time" in args:
+        args["time"] += d_time
+    else:
+        args["start"] += d_time
+        args["end"] += d_time
+    return args
 
-    def test_single(self, marker_tlui, tilia_state):
-        tilia_state.duration = 100
-        marker_tlui.create_marker(50)
+
+@pytest.mark.parametrize(
+    "tlui, args",
+    [
+        (
+            "audiowave_tlui",
+            {"kind": ComponentKind.AUDIOWAVE, "start": 50, "end": 55, "amplitude": 10},
+        ),
+        ("beat_tlui", {"kind": ComponentKind.BEAT, "time": 50}),
+        ("harmony_tlui", {"kind": ComponentKind.HARMONY, "time": 50}),
+        ("harmony_tlui", {"kind": ComponentKind.MODE, "time": 50}),
+        (
+            "hierarchy_tlui",
+            {"kind": ComponentKind.HIERARCHY, "start": 50, "end": 55, "level": 2},
+        ),
+        ("marker_tlui", {"kind": ComponentKind.MARKER, "time": 50}),
+        ("pdf_tlui", {"kind": ComponentKind.PDF_MARKER, "time": 50, "page_number": 1}),
+        # "score_tlui", setup is a little too complicated
+        # "slider_tlui", not clickable
+    ],
+)
+class TestModifierSelect:
+    @pytest.mark.parametrize(
+        "modifiers,changes_time,select_value",
+        [
+            (["alt", "ctrl"], False, [True, False]),
+            (["alt"], False, [True, True]),
+            (["ctrl"], True, [True, False]),
+        ],
+        ids=["alt+ctrl", "alt", "ctrl"],
+    )
+    def test_single(
+        self,
+        qtui,
+        tlui,
+        args,
+        tilia_state,
+        request,
+        modifiers,
+        changes_time,
+        select_value,
+    ):
+        tlui = request.getfixturevalue(tlui)
+        tlui.create_component(**args)
+
+        if "time" in args:
+            click_time = tlui[0].get_data("time")
+        else:
+            click_time = tlui[0].get_data("start")
 
         for i in range(10):
-            click_timeline_ui(
-                marker_tlui, marker_tlui[0].get_data("time"), modifier=["alt", "ctrl"]
-            )
-            if i % 2 == 0:
-                assert marker_tlui[0].is_selected()
-            else:
-                assert not marker_tlui[0].is_selected()
+            click_timeline_ui(tlui, click_time, modifier=modifiers)
+            assert tlui[0].is_selected() == select_value[i % 2]
+            assert tilia_state.current_time == (click_time if changes_time else 0)
 
-        # assert we have not seeked
-        assert tilia_state.current_time == 0
+    @pytest.mark.parametrize(
+        "modifiers,changes_time,select_value",
+        [
+            (
+                ["alt", "ctrl"],
+                False,
+                [(True, False), (True, True), (False, True), (False, False)],
+            ),
+            (
+                ["alt"],
+                False,
+                [(True, False), (False, True), (True, False), (False, True)],
+            ),
+            (
+                [
+                    "ctrl",
+                    True,
+                    [(True, False), (True, True), (False, True), (False, False)],
+                ]
+            ),
+        ],
[…]
