### 2025-04-09 7401de6
test: include score_tl in parametrised tests
