### 2025-04-11 d479690
fix: create method to click directly on a ui element
