### 2025-04-11 8255061
fix: create method to click directly on a ui element
