### 2025-04-11 6b6ece5
refactor: remove unnecessary enum
