### 2025-04-11 33710ee
refactor: remove unnecessary enum
