### 2025-04-11 c405372
fix: updated help link
