### 2025-04-12 ca288c2
fix: None returned on dialog cancel was not handled
