### 2025-04-12 991f88b
fix: dialog header
