### 2025-05-27 #7893 @FelipeDefensor
> [@FelipeDefensor](https://github.com/FelipeDefensor) – could you take a look a [this list of potential reviewers](https://reviewers.joss.theoj.org/lookup) and identify a few people who would be good candidates to review this submission?

@arfon, I wasn't able to find any reviewer that seemed like a particularly good match in the list. Are we restricted to those? If not, here are some suggestions @azfoo has made earlier, three of them being former JOSS authors/reviewers.

> Regarding reviewers, Maik Köster (m-k94) is fairly familiar with the app from the perspective of a user -- we have no prior collaboration but are in the midst of planning a paper on dataset alignment, tangentially related to the work of this software.
Other potential reviewers are Stefan Balke (stefan-balke), Magdalena Fuentes (magdalenafuentes) and Brian McFee (bmcfee).
