### 2025-06-10 339ddc3
dependency: upgrade setuptools due to vulnerability
