+++ requirements.txt
@@ -15,7 +15,5 @@ pytube~=15.0.0
 pytest>=8.0.0
 pytest-env>=1.1.5
 icecream~=2.1.3
-requests~=2.32.3
-setuptools~=72.2.0
 colorama~=0.4.6
 sentry-sdk>=2.21.0
