### 2025-06-10 d0e0c92
version: 0.5.12
