### 2025-06-15 #7893 @arfon
@lutzhamel @sdruskat – 👋 would any of you be willing to review this submission for JOSS? The submission under consideration is TiLiA: a timeline annotator

Many thanks
Arfon
