### 2025-06-17 #7893 @arfon
Thanks @expectopatronum! Very happy to be flexible on timing here (and grateful you're willing to do both ✨)
