### 2025-06-17 #8417 @editorialbot
[REVIEW]: TiLiA: a timeline annotator
**Submitting author:** @FelipeDefensor (<a href="http://orcid.org/0000-0002-9965-042X">Martins, Felipe D.</a>)
**Repository:** https://github.com/TimeLineAnnotator/desktop
**Branch with paper.md** (empty if default branch): dev
**Version:** v0.5.9
