### 2025-06-19 6073a02
fix: do not crash when asking for invalid fraction value
This happens when importing the `score_xml_unfolded/Beethoven_Op031No2-01.xml` from the Beethoven Piano Sonatas Dataset.

Seems harmless enough to return the closest fraction, but I have not investigated this in any depth.
