### 2025-06-19 632dfee
fix: cli commands with trailing whitespaces
We need to strip the whitespaces, otherwise they error.
