### 2025-06-19 4a47cd9
feat: colored cli output
This makes script output much more readable and easy to debug.
