### 2025-06-19 63d71b3
feat: colored cli output
This makes script output much more readable and easy to debug.
