### 2025-06-19 28740fa
feat: cli command - clear
