### 2025-06-19 f3e1ca5
feat: cli command - clear
