+++ tilia/ui/cli/ui.py
@@ -24,6 +24,7 @@ from tilia.ui.cli import (
     generate_scripts,
     open,
     export,
+    clear,
 )
 from tilia.ui.cli.io import ask_yes_or_no
 from tilia.ui.cli.player import CLIVideoPlayer, CLIYoutubePlayer
@@ -56,6 +57,7 @@ class CLI:
         script.setup_parser(self.subparsers, self.parse_and_run)
         open.setup_parser(self.subparsers)
         export.setup_parser(self.subparsers)
+        clear.setup_parser(self.subparsers)
 
     @staticmethod
     def parse_command(arg_string):
