### 2025-06-20 74a364e
fix: import score command & remove unnecessary "csv" arg
