### 2025-06-20 4890fa3
feat: estimate duration of last beat when importing
