### 2025-06-20 bccb528
feat: accept .xml extension in imported scores
