### 2025-07-03 2609197
chore: type hint
