### 2025-07-03 73303b8
fix: crash when inputting harmony extensions
This only fixes the crash. Parsing extensions must be implemented separately.
