### 2025-07-04 e4fe45e
feat: CLI metadata set - add field if it doesn't exist
