### 2025-07-04 e7198d9
feat: CLI metadata set - add field if it doesn't exist
