### 2025-07-09 2d52486
fix: also skip parsing attributes of implicit measures
