### 2025-07-09 ccb6e48
fix: rounding error leading to out-of-bounds notes
