### 2025-07-10 50bbd8e
fix: add time validation to score timeline components
