### 2025-07-10 5bc8573
fix: add time validation to score timeline components
