### 2025-07-14 702ecd2
feat: CLI script command - "encoding" argument
