### 2025-07-15 86b1e4f
refactor: always use Post.DISPLAY_ERROR
