### 2025-07-15 6b037cf
refactor: always use Post.DISPLAY_ERROR
