### 2025-07-16 TimeLineAnnotator/desktop#358 @expectopatronum
Questions regarding the paper
Hi @FelipeDefensor, I am looking forward to [review](https://github.com/openjournals/joss-reviews/issues/8417) your application!

I hope it is ok that I opened an issue for questions regarding the paper. I might add new comments with questions while I am reading/reviewing the paper.

Tiny question regarding contributing authors: the names/people do not exactly match, e.g. I don't see Mark Gotham among the committing authors, but there are a few names that are not on the paper - is this correct? Just asking because I need to tick `Does the full list of paper authors seem appropriate and complete?`

Thanks & best regards
Verena
