### 2025-07-16 TimeLineAnnotator/desktop#358 @FelipeDefensor
Hello, Verena.

Please do open as many issues as you see fit :).

I imagined the author list would come up at some point. Here is the commit count, for easier reference:

   961	jiraffarij
   399	azfoo
   288	Felipe Martins
    41	Anne Foo
    32	Felipe Defensor
     3	Paper Workflow Bot
     3	felipe.martins
     1	lgabech

`jiraffarij`, `Felipe Martins` and `Felipe Defensor` are all me. `Anne Foo` is also `azfoo`. Mark has made non-code contributions only. I realize now that we listed all authors as equal contributors in the submission. Perhaps better would be: 1. Felipe M., 2. Anne F. and 3. Mark G..
