### 2025-07-20 69bfd8b
refactor: merge function
