### 2025-07-20 f4d8f03
feat: return result from post request
