### 2025-07-20 b31fa5f
refactor: return success from load media functions
