+++ tilia/media/player/youtube.py
@@ -117,16 +117,24 @@ class YouTubePlayer(Player):
         end: float = 0.0,
         initial_duration: float = 0.0,
     ):
+        """
+        Returns True if media loading has *started* succesfully, False otherwise.
+        Loading is asynchronous, and self.on_media_load_done will be called
+         when it is completed.
+        If initial_duration is provided, it will be available when returning.
+        """
         if not self.view.isVisible():
             self.view.show()
 
-        super().load_media(path, start, end)
+        success = super().load_media(path, start, end)
 
         if initial_duration:
             # This ensures duration is available
             # after opening a file, as self.on_media_duration_available
             # will not be called immediately by the engine.
             self.on_media_duration_available(initial_duration)
+
+        return success
 
     def on_media_load_done(self, path, start, end):
         self.media_path = str(path)
