### 2025-07-20 1351d4e
refactor: return success from load media functions
