### 2025-07-20 a4aebdc
feat: get YouTube video duration when loading from CLI
