+++ tests/ui/cli/test_load_media.py
@@ -1,7 +1,9 @@
+import os
 from unittest.mock import patch
 import shutil
 
 import pytest
+import httpx
 
 from tests.constants import (
     EXAMPLE_MEDIA_PATH,
@@ -9,6 +11,7 @@ from tests.constants import (
     EXAMPLE_MEDIA_SCALE_FACTOR,
 )
 from tilia.requests import get, Get
+from tilia.ui.cli.player import CLIYoutubePlayer
 
 
 def test_load_media(cli, tilia_state):
@@ -103,3 +106,126 @@ def test_with_timelines_scale_not_provied_answer_crop(
     assert tilia_state.duration == EXAMPLE_MEDIA_DURATION
     assert len(marker_tl) == 1
     assert marker_tl[0].get_data("time") == 5
+
+
+@pytest.mark.skipif(
+    os.environ.get("YOUTUBE_API_KEY") is None,
+    reason="YOUTUBE_API_KEY environment variable is required to run this test.",
+)
+class TestLoadYoutube:
+    PREFIX = "https://www.youtube.com/watch?v="
+    EXAMPLE_VIDEO_ID = "WDzeHNpNaoM"
+    EXAMPLE_VIDEO_DURATION = 3056
+    INITIAL_DURATION = 1
+
+    @pytest.fixture(autouse=True)
+    def set_duration(self, tilia_state):
+        tilia_state.duration = self.INITIAL_DURATION
+
+    @staticmethod
+    def skip_if_no_youtube_api_key(func):
+        return
+
+    def load_media(self, cli, id):
+        cli.parse_and_run(f"load-media {self.PREFIX}{id}")
+
+    def assert_error(self, tilia_errors, message, tilia_state, prev_duration):
+        tilia_errors.assert_error()
+        tilia_errors.assert_in_error_message(message)
+        assert tilia_state.duration == prev_duration
+
+    def test_load(self, cli, tilia_errors):
+        self.load_media(cli, self.EXAMPLE_VIDEO_ID)
+
+        tilia_errors.assert_no_error()
+        assert get(Get.MEDIA_DURATION) == self.EXAMPLE_VIDEO_DURATION
+
+    def test_load_youtu_dot_be(self, cli, tilia_errors):
+        cli.parse_and_run(
+            f"load-media https://youtu.be/watch?v={self.EXAMPLE_VIDEO_ID}"
+        )
+
+        tilia_errors.assert_no_error()
+        assert get(Get.MEDIA_DURATION) == self.EXAMPLE_VIDEO_DURATION
+
+    def test_load_from_playlist(self, cli, tilia_errors):
+        cli.parse_and_run(
+            'load-media "https://www.youtube.com/watch?v=adLGHcj_fmA&list=RDQMgEzdN5RuCXE&index=2"'
+        )
+
+        tilia_errors.assert_no_error()
+        assert get(Get.MEDIA_DURATION) == 249
+
+    @patch("httpx.get")
+    def test_no_connection(self, mock_httpx, cli, tilia_errors, tilia_state):
+        mock_httpx.side_effect = httpx.RequestError("")
+
+        self.load_media(cli, self.EXAMPLE_VIDEO_ID)
+
+        self.assert_error(
+            tilia_errors,
+            CLIYoutubePlayer.CONNECTION_ERROR_MESSAGE,
+            tilia_state,
+            self.INITIAL_DURATION,
+        )
+
+    @patch("httpx.Response.raise_for_status")
+    def test_request_failed(self, mock_httpx, cli, tilia_errors, tilia_state):
+        mock_httpx.side_effect = httpx.HTTPStatusError(
+            "",
+            request=httpx.Request("", self.PREFIX + self.EXAMPLE_VIDEO_ID),
[…]
