+++ requirements.txt
@@ -10,7 +10,8 @@ PyQt6-WebEngine==6.6.0
 PyQt6-WebEngine-Qt6==6.6.0
 python-dotenv>=1.0.1
 pypdf~=4.2.0
-pytube~=15.0.0
+httpx~=0.28.1
+isodate~=0.7.2
 
 pytest>=8.0.0
 pytest-env>=1.1.5
