### 2025-07-20 808955e
feat: get YouTube video duration when loading from CLI
