+++ setup.cfg
@@ -19,8 +19,9 @@ install_requires =
     PyQt6-WebEngine-Qt6==6.6.0
     python-dotenv>=1.0.1
     pypdf~=4.2.0
-    pytube~=15.0.0
     sentry-sdk>=2.21.0
+    httpx~=0.28.1
+    isodate~=0.7.2
 python_requires = >=3.10
 zip_safe = no
 
