### 2025-07-20 2ff0e59
fix: CLI commands with trailing spaces
