### 2025-07-20 d47f740
fix: CLI commands with trailing spaces
