### 2025-07-20 6c04b1a
feat: comments in CLI scripts
