### 2025-07-20 d4ef2ba
feat: comments in CLI scripts
