### 2025-07-20 8113932
test: --encoding CLI option
