### 2025-07-20 e15aeba
test: CLI "clear" command
