### 2025-07-20 7a60748
refactor: don't use exceptions for flow control
