### 2025-07-20 2b2e54a
test: import score with CLI
