### 2025-07-20 e0573d7
feat: estimate duration of last measure when importing
