### 2025-07-20 0fcc6c6
feat: estimate duration of last measure when importing
