### 2025-07-20 4c763c7
wip!: remove ordinal funcs from args.py
