+++ tilia/ui/timelines/collection/requests/timeline.py
@@ -21,12 +21,6 @@ request_to_scope: dict[Post, TlRequestSelector] = {
     ),
     Post.TIMELINE_NAME_SET: TlRequestSelector(NOT_SLIDER, TimelineSelector.FIRST),
     Post.TIMELINE_HEIGHT_SET: TlRequestSelector(NOT_SLIDER, TimelineSelector.FIRST),
-    Post.TIMELINE_ORDINAL_INCREASE_FROM_MANAGE_TIMELINES: TlRequestSelector(
-        ALL, TimelineSelector.FROM_MANAGE_TIMELINES_TO_PERMUTE
-    ),
-    Post.TIMELINE_ORDINAL_DECREASE_FROM_MANAGE_TIMELINES: TlRequestSelector(
-        ALL, TimelineSelector.FROM_MANAGE_TIMELINES_TO_PERMUTE
-    ),
     Post.TIMELINE_IS_VISIBLE_SET_FROM_MANAGE_TIMELINES: TlRequestSelector(
         ALL, TimelineSelector.FROM_MANAGE_TIMELINES_CURRENT
     ),
@@ -38,11 +32,5 @@ request_to_scope: dict[Post, TlRequestSelector] = {
     ),
     Post.TIMELINE_DELETE_FROM_CONTEXT_MENU: TlRequestSelector(
         ALL, TimelineSelector.FROM_CONTEXT_MENU
-    ),
-    Post.TIMELINE_ORDINAL_DECREASE_FROM_CONTEXT_MENU: TlRequestSelector(
-        ALL, TimelineSelector.FROM_CONTEXT_MENU_TO_PERMUTE
-    ),
-    Post.TIMELINE_ORDINAL_INCREASE_FROM_CONTEXT_MENU: TlRequestSelector(
-        ALL, TimelineSelector.FROM_CONTEXT_MENU_TO_PERMUTE
     ),
 }
