+++ tilia/ui/timelines/collection/requests/timeline_uis.py
@@ -1,3 +1,11 @@
 from tilia.requests import Post
 
-requests = [Post.TIMELINES_CLEAR, Post.TIMELINE_ADD, Post.BEAT_TIMELINE_FILL]
+requests = [
+    Post.TIMELINES_CLEAR,
+    Post.TIMELINE_ADD,
+    Post.BEAT_TIMELINE_FILL,
+    Post.TIMELINE_ORDINAL_INCREASE_FROM_MANAGE_TIMELINES,
+    Post.TIMELINE_ORDINAL_DECREASE_FROM_MANAGE_TIMELINES,
+    Post.TIMELINE_ORDINAL_DECREASE_FROM_CONTEXT_MENU,
+    Post.TIMELINE_ORDINAL_INCREASE_FROM_CONTEXT_MENU,
+]
