### 2025-07-20 9bdecae
wip!: remove timeline set name func from args.py
