### 2025-07-20 f7e3548
wip!: remove timeline set height func from args.py
