### 2025-07-20 a0f52c4
wip!: remove timeline is visible set func from args.py
