+++ tilia/ui/timelines/collection/requests/timeline_uis.py
@@ -8,4 +8,5 @@ requests = [
     Post.TIMELINE_ORDINAL_DECREASE_FROM_MANAGE_TIMELINES,
     Post.TIMELINE_ORDINAL_DECREASE_FROM_CONTEXT_MENU,
     Post.TIMELINE_ORDINAL_INCREASE_FROM_CONTEXT_MENU,
+    Post.TIMELINE_DELETE_FROM_MANAGE_TIMELINES,
 ]
