### 2025-07-20 428c35f
wip!: remove add pre start post end set funcs from args.py
