### 2025-07-20 a1c6312
wip!: remove set element color args from args.py NOW YOU CAN REMOVE THE WHOLE MODULE!!
