+++ tilia/ui/timelines/score/request_handlers.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING
 
-from tilia.requests import Post
+from tilia.requests import Post, get, Get
 from tilia.ui.timelines.base.request_handlers import ElementRequestHandler
 
 if TYPE_CHECKING:
@@ -21,8 +21,12 @@ class ScoreTimelineUIElementRequestHandler(ElementRequestHandler):
             },
         )
 
-    def on_color_set(self, elements, value, **_):
-        self.timeline_ui.set_elements_attr(elements, "color", value.name())
+    def on_color_set(self, elements, **_):
+        success, color = get(Get.FROM_USER_COLOR)
+        if not success or not color.isValid():
+            return False
+
+        self.timeline_ui.set_elements_attr(elements, "color", color.name())
         return True
 
     def on_color_reset(self, elements, *_, **__):
