+++ setup.cfg
@@ -22,6 +22,7 @@ install_requires =
     sentry-sdk>=2.21.0
     httpx~=0.28.1
     isodate~=0.7.2
+    colorama~=0.4.6
 python_requires = >=3.10
 zip_safe = no
 
