+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = TiLiA
-version = 0.5.13
+version = 0.5.14
 description = A GUI for creating and visualizing annotations over audio and video files.
 author = Felipe Defensor et al.
 
