### 2025-07-21 77d588e
wip!: merge request_to_scope dicts
