### 2025-07-21 f9dd6c8
wip!: merge ORDINAL_INCREASE and ORDINAL_DECREASE requests
