+++ tilia/requests/post.py
@@ -158,10 +158,8 @@ class Post(Enum):
     TIMELINE_KIND_INSTANCED = auto()
     TIMELINE_KIND_NOT_INSTANCED = auto()
     TIMELINE_NAME_SET = auto()
-    TIMELINE_ORDINAL_DECREASE_FROM_MANAGE_TIMELINES = auto()
-    TIMELINE_ORDINAL_INCREASE_FROM_MANAGE_TIMELINES = auto()
-    TIMELINE_ORDINAL_DECREASE_FROM_CONTEXT_MENU = auto()
-    TIMELINE_ORDINAL_INCREASE_FROM_CONTEXT_MENU = auto()
+    TIMELINE_ORDINAL_PERMUTE_FROM_MANAGE_TIMELINES = auto()
+    TIMELINE_ORDINAL_PERMUTE_FROM_CONTEXT_MENU = auto()
     TIMELINE_ORDINAL_SET = auto()
     TIMELINE_PREPARING_TO_DRAG = auto()
     TIMELINE_SET_DATA_DONE = auto()
