+++ tilia/ui/timelines/collection/collection.py
@@ -207,10 +207,10 @@ class TimelineUIs:
             Post.TIMELINES_CLEAR,
             Post.TIMELINE_ADD,
             Post.BEAT_TIMELINE_FILL,
-            Post.TIMELINE_ORDINAL_INCREASE_FROM_MANAGE_TIMELINES,
-            Post.TIMELINE_ORDINAL_DECREASE_FROM_MANAGE_TIMELINES,
-            Post.TIMELINE_ORDINAL_DECREASE_FROM_CONTEXT_MENU,
-            Post.TIMELINE_ORDINAL_INCREASE_FROM_CONTEXT_MENU,
+            Post.TIMELINE_ORDINAL_PERMUTE_FROM_MANAGE_TIMELINES,
+            Post.TIMELINE_ORDINAL_PERMUTE_FROM_MANAGE_TIMELINES,
+            Post.TIMELINE_ORDINAL_PERMUTE_FROM_CONTEXT_MENU,
+            Post.TIMELINE_ORDINAL_PERMUTE_FROM_CONTEXT_MENU,
             Post.TIMELINE_DELETE_FROM_MANAGE_TIMELINES,
         ]:
             listen(
