+++ tests/ui/timelines/harmony/test_harmony_timeline_ui.py
@@ -4,7 +4,6 @@ import tilia.ui.actions
 from tests.mock import Serve
 from tests.ui.timelines.harmony.interact import click_harmony_ui
 from tilia.requests import Get
-from tilia.ui.actions import TiliaAction
 
 FLAT_SIGN = "`b"
 SHARP_SIGN = "`#"
@@ -22,7 +21,7 @@ def add_harmony(**kwargs):
     }
     default_params.update(kwargs)
     with Serve(Get.FROM_USER_HARMONY_PARAMS, (True, default_params)):
-        tilia.ui.actions.trigger(TiliaAction.HARMONY_ADD)
+        tilia.ui.actions.trigger("harmony_add")
 
 
 def add_mode(**kwargs):
@@ -34,7 +33,7 @@ def add_mode(**kwargs):
     }
     default_params.update(kwargs)
     with Serve(Get.FROM_USER_MODE_PARAMS, (True, default_params)):
-        tilia.ui.actions.trigger(TiliaAction.MODE_ADD)
+        tilia.ui.actions.trigger("mode_add")
 
 
 class TestRomanNumeralDisplay:
@@ -136,9 +135,9 @@ class TestCopyPaste:
 
         click_harmony_ui(harmony_tlui.modes()[0])
         click_harmony_ui(harmony_tlui.harmonies()[1], modifier="ctrl")
-        user_actions.trigger(TiliaAction.TIMELINE_ELEMENT_COPY)
+        user_actions.trigger("timeline_element_copy")
 
         click_harmony_ui(harmony_tlui.harmonies()[1])
-        user_actions.trigger(TiliaAction.TIMELINE_ELEMENT_PASTE)
+        user_actions.trigger("timeline_element_paste")
 
         assert len(harmony_tlui) == 5
