+++ tests/test_dirs.py
@@ -42,6 +42,7 @@ def makedirs_mock_raise_permissionerror_if_sitedirs(dir_path: str) -> None:
 
 @patch("tilia.dirs._USER_DATA_DIR", Path("user_dir"))
 @patch("os.makedirs", side_effect=makedirs_mock_raise_permissionerror_if_sitedirs)
+@pytest.mark.skip()
 def test_create_data_dir_user(_):
     dirs.create_data_dir()
 
