+++ tests/timelines/beat/test_beat_timeline.py
@@ -1,19 +1,17 @@
 import pytest
 
-from tilia.ui.actions import TiliaAction
-
 
 class TestBeatTimeline:
     def test_create_beat_at_same_time_fails(self, beat_tlui, user_actions):
-        user_actions.trigger(TiliaAction.BEAT_ADD)
-        user_actions.trigger(TiliaAction.BEAT_ADD)
+        user_actions.trigger("beat_add")
+        user_actions.trigger("beat_add")
         assert len(beat_tlui) == 1
 
     def test_create_beat_at_negative_time_fails(
         self, beat_tlui, tilia_state, user_actions
     ):
         tilia_state.current_time = -10
-        user_actions.trigger(TiliaAction.BEAT_ADD)
+        user_actions.trigger("beat_add")
         assert len(beat_tlui) == 0
 
     def test_create_beat_at_time_bigger_than_media_duration_fails(
@@ -21,7 +19,7 @@ class TestBeatTimeline:
     ):
         tilia_state.duration = 100
         tilia_state.current_time = 101
-        user_actions.trigger(TiliaAction.BEAT_ADD)
+        user_actions.trigger("beat_add")
         assert len(beat_tlui) == 0
 
     def test_create_beat_at_middle_updates_next_beats_is_first_in_measure(
@@ -29,14 +27,14 @@ class TestBeatTimeline:
     ):
         beat_tlui.timeline.beat_pattern = [2]
         tilia_state.current_time = 0
-        user_actions.trigger(TiliaAction.BEAT_ADD)
+        user_actions.trigger("beat_add")
         tilia_state.current_time = 10
-        user_actions.trigger(TiliaAction.BEAT_ADD)
+        user_actions.trigger("beat_add")
         tilia_state.current_time = 20
-        user_actions.trigger(TiliaAction.BEAT_ADD)
+        user_actions.trigger("beat_add")
 
         tilia_state.current_time = 5
-        user_actions.trigger(TiliaAction.BEAT_ADD)
+        user_actions.trigger("beat_add")
 
         assert beat_tlui[0].get_data("is_first_in_measure") is True
         assert beat_tlui[1].get_data("is_first_in_measure") is False
