+++ tilia/ui/timelines/beat/toolbar.py
@@ -1,13 +1,12 @@
 from __future__ import annotations
 
-from tilia.ui.actions import TiliaAction
 from tilia.ui.timelines.toolbar import TimelineToolbar
 
 
 class BeatTimelineToolbar(TimelineToolbar):
     ACTIONS = [
-        TiliaAction.BEAT_ADD,
-        TiliaAction.BEAT_DISTRIBUTE,
-        TiliaAction.BEAT_SET_MEASURE_NUMBER,
-        TiliaAction.BEAT_RESET_MEASURE_NUMBER,
+        "beat_add",
+        "beat_distribute",
+        "beat_set_measure_number",
+        "beat_reset_measure_number",
     ]
