### 2025-07-26 6eb53c8
refactor: hardcoded values
