### 2025-07-26 b20f768
!wip: user actions.register for timeline.add commands
