+++ tests/ui/timelines/test_timelineui_collection.py
@@ -18,17 +18,17 @@ from tilia.ui import actions
 from tilia.ui.enums import ScrollType
 from tilia.ui.timelines.base.request_handlers import TimelineRequestHandler
 from tilia.ui.timelines.base.timeline import TimelineUI
-from tilia.ui.timelines.collection.request_handler import TimelineUIsRequestHandler
+from tilia.ui.timelines.collection.collection import TimelineUIs
 from tilia.ui.timelines.marker.request_handlers import MarkerUIRequestHandler
 from tilia.ui.coords import time_x_converter
 from tilia.ui.dialogs.add_timeline_without_media import AddTimelineWithoutMedia
 
 ADD_TIMELINE_ACTIONS = [
-    "timelines_add_hierarchy_timeline",
-    "timelines_add_beat_timeline",
-    "timelines_add_harmony_timeline",
-    "timelines_add_marker_timeline",
-    "timelines_add_audiowave_timeline",
+    "timelines.add.hierarchy",
+    "timelines.add.beat",
+    "timelines.add.harmony",
+    "timelines.add.marker",
+    "timelines.add.audiowave",
 ]
 
 
@@ -44,11 +44,11 @@ class TestTimelineUICreation:
 
     def test_create_multiple(self, tilia_state, tluis, user_actions):
         create_actions = [
-            "timelines_add_harmony_timeline",
-            "timelines_add_marker_timeline",
-            "timelines_add_beat_timeline",
-            "timelines_add_hierarchy_timeline",
-            "timelines_add_audiowave_timeline",
+            "timelines.add.harmony",
+            "timelines.add.marker",
+            "timelines.add.beat",
+            "timelines.add.hierarchy",
+            "timelines.add.audiowave",
         ]
         with (
             Serve(Get.FROM_USER_BEAT_PATTERN, (True, [1])),
@@ -68,7 +68,7 @@ class TestTimelineUICreation:
         ):
             with Serve(Get.FROM_USER_FLOAT, (True, 10)):
                 with Serve(Get.FROM_USER_STRING, (True, "")):
-                    user_actions.trigger("timelines_add_marker_timeline")
+                    user_actions.trigger("timelines.add.marker")
         assert tilia_state.duration == 10
         assert len(tluis) == 1
 
@@ -82,7 +82,7 @@ class TestTimelineUICreation:
         ):
             with Serve(Get.FROM_USER_MEDIA_PATH, (True, EXAMPLE_MEDIA_PATH)):
                 with Serve(Get.FROM_USER_STRING, (True, "")):
-                    user_actions.trigger("timelines_add_marker_timeline")
+                    user_actions.trigger("timelines.add.marker")
         assert tilia_state.duration == EXAMPLE_MEDIA_DURATION
         assert len(tluis) == 1
 
@@ -95,7 +95,7 @@ class TestTimelineUICreation:
             (True, AddTimelineWithoutMedia.Result.SET_DURATION),
         ):
             with Serve(Get.FROM_USER_FLOAT, (False, 10)):
-                user_actions.trigger("timelines_add_marker_timeline")
+                user_actions.trigger("timelines.add.marker")
         assert tilia_state.duration == 0
         assert len(tluis) == 0
 
@@ -108,7 +108,7 @@ class TestTimelineUICreation:
             (True, AddTimelineWithoutMedia.Result.LOAD_MEDIA),
         ):
             with Serve(Get.FROM_USER_MEDIA_PATH, (False, EXAMPLE_MEDIA_PATH)):
-                user_actions.trigger("timelines_add_marker_timeline")
+                user_actions.trigger("timelines.add.marker")
         assert tilia_state.duration == 0
         assert len(tluis) == 0
 
@@ -120,7 +120,7 @@ class TestTimelineUICreation:
 
[…]
