+++ tests/ui/cli/timelines/test_cli_timelines.py
@@ -52,7 +52,7 @@ class TestTimelineRemove:
 
     def test_by_name_one_timeline(self, cli, tls, user_actions):
         with Serve(Get.FROM_USER_STRING, (True, "test")):
-            user_actions.trigger("timelines_add_hierarchy_timeline")
+            user_actions.trigger("timelines.add.hierarchy")
 
         cli.parse_and_run("timeline remove name test")
 
@@ -72,7 +72,7 @@ class TestTimelineRemove:
 
     def test_remove_by_name_not_found(self, cli, tls, user_actions):
         with Serve(Get.FROM_USER_STRING, (True, "test")):
-            user_actions.trigger("timelines_add_hierarchy_timeline")
+            user_actions.trigger("timelines.add.hierarchy")
 
         with patch("builtins.print") as mock_print:
             cli.parse_and_run("timeline remove name othername")
