### 2025-07-27 5d80ba6
!wip: rename commands.trigger to commands.execute
