+++ tests/ui/timelines/hierarchy/test_hierarchy_timeline_ui.py
@@ -42,7 +42,7 @@ class TestActions:
         tlui.create_hierarchy(3, 4, 1)
 
         tlui.select_element(tlui[0])
-        user_actions.trigger("hierarchy_increase_level")
+        user_actions.execute("hierarchy_increase_level")
 
         assert tlui[2].get_data("level") == 2
         assert tlui[2].get_data("start") == 0
@@ -58,7 +58,7 @@ class TestActions:
         tlui.select_element(tlui[0])
         tlui.select_element(tlui[1])
         tlui.select_element(tlui[2])
-        user_actions.trigger("hierarchy_increase_level")
+        user_actions.execute("hierarchy_increase_level")
 
         assert tlui[0].get_data("level") == 2
         assert tlui[1].get_data("level") == 2
@@ -70,7 +70,7 @@ class TestActions:
         tlui.create_hierarchy(3, 4, 2)
 
         tlui.select_element(tlui[0])
-        user_actions.trigger("hierarchy_decrease_level")
+        user_actions.execute("hierarchy_decrease_level")
 
         assert tlui[0].get_data("level") == 1
         assert tlui[1].get_data("level") == 2
@@ -84,7 +84,7 @@ class TestActions:
         tlui.select_element(tlui[0])
         tlui.select_element(tlui[1])
         tlui.select_element(tlui[2])
-        user_actions.trigger("hierarchy_decrease_level")
+        user_actions.execute("hierarchy_decrease_level")
 
         assert tlui[0].get_data("level") == 1
         assert tlui[1].get_data("level") == 1
@@ -95,7 +95,7 @@ class TestActions:
         tlui.select_element(tlui[0])
 
         with Serve(Get.FROM_USER_COLOR, (True, QColor("#000"))):
-            user_actions.trigger("timeline_element_color_set")
+            user_actions.execute("timeline_element_color_set")
 
         assert tlui[0].get_data("color") == "#000000"
 
@@ -104,9 +104,9 @@ class TestActions:
         tlui.select_element(tlui[0])
 
         with Serve(Get.FROM_USER_COLOR, (True, QColor("#000"))):
-            user_actions.trigger("timeline_element_color_set")
+            user_actions.execute("timeline_element_color_set")
 
-        user_actions.trigger("timeline_element_color_reset")
+        user_actions.execute("timeline_element_color_reset")
 
         assert tlui[0].get_data("color") is None
 
@@ -115,7 +115,7 @@ class TestActions:
         tlui.select_element(tlui[0])
 
         with Serve(Get.FROM_USER_FLOAT, (True, 0.1)):
-            user_actions.trigger("hierarchy_add_pre_start")
+            user_actions.execute("hierarchy_add_pre_start")
 
         assert tlui[0].get_data("pre_start") != tlui[0].get_data("start")
         assert tlui[0].pre_start_handle
@@ -125,7 +125,7 @@ class TestActions:
         tlui.select_element(tlui[0])
 
         with Serve(Get.FROM_USER_FLOAT, (True, 0.1)):
-            user_actions.trigger("hierarchy_add_post_end")
+            user_actions.execute("hierarchy_add_post_end")
 
         assert tlui[0].get_data("post_end") != tlui[0].get_data("end")
         assert tlui[0].post_end_handle
@@ -134,7 +134,7 @@ class TestActions:
         tlui.create_hierarchy(0, 1, 1)
         assert len(tlui) == 1
         tilia_state.current_time = 0.5
-        user_actions.trigger("hierarchy_split")
+        user_actions.execute("hierarchy_split")
 
         assert len(tlui) == 2
 
@@ -145,7 +145,7 @@ class TestActions:
         tlui.select_element(tlui[0])
         tlui.select_element(tlui[1])
 
[…]
