+++ tests/ui/timelines/slider/test_slider_timeline_ui.py
@@ -6,7 +6,7 @@ def test_undo_redo(slider_tlui, marker_tlui, user_actions):
     post(Post.APP_RECORD_STATE, "test state")
 
     # using marker tl to trigger an actions that can be undone
-    user_actions.trigger("marker_add")
+    user_actions.execute("marker_add")
 
     post(Post.EDIT_UNDO)
     assert len(marker_tlui) == 0
