### 2025-07-27 b461b84
!wip: return results from commands
