### 2025-07-27 cc1801e
!wip: register import commands
