+++ tests/parsers/csv/test_markers_from_csv.py
@@ -8,17 +8,11 @@ from PyQt6.QtWidgets import QFileDialog
 from tests.parsers.csv.common import assert_in_errors
 from tilia.ui import commands
 from tilia.ui.format import format_media_time
-from tilia.ui.ui_import import on_import_from_csv
 
 
 def patch_import(by: Literal["time", "measure"], tl, data) -> tuple[str, list[str]]:
     status = ""
     errors = []
-
-    def mock_import(timeline_uis, tlkind):
-        nonlocal status, errors
-        status, errors = on_import_from_csv(timeline_uis, tlkind)
-        return status, errors
 
     with (
         patch(
@@ -27,10 +21,9 @@ def patch_import(by: Literal["time", "measure"], tl, data) -> tuple[str, list[st
         ),
         patch.object(QFileDialog, "exec", return_value=True),
         patch.object(QFileDialog, "selectedFiles", return_value=[Path()]),
-        patch("tilia.ui.qtui.on_import_from_csv", side_effect=mock_import),
         patch("builtins.open", mock_open(read_data=data)),
     ):
-        commands.execute("import_csv_marker_timeline")
+        status, errors = commands.execute("timelines.import.marker")
     return status, errors
 
 
