### 2025-07-27 44acafa
!wip: use callback as required argument in register_action
