### 2025-07-27 e6fd4d8
!wip: rename register_action to register_command
