+++ tilia/ui/windows/__init__.py
@@ -2,4 +2,5 @@ from tilia.ui.windows.kinds import WindowKind
 
 
 class UniqueWindowDuplicate(Exception):
-    def __init__(self, window_kind: WindowKind): ...
+    def __init__(self, window_kind: WindowKind):
+        ...
