+++ tilia/ui/timelines/base/element.py
@@ -64,7 +64,8 @@ class TimelineUIElement(ABC):
         return self in self.timeline_ui.selected_elements
 
     @abstractmethod
-    def child_items(self): ...
+    def child_items(self):
+        ...
 
     def selection_triggers(self):
         return self.child_items()
@@ -85,9 +86,11 @@ class TimelineUIElement(ABC):
         menu = self.CONTEXT_MENU_CLASS(self)
         menu.exec(QPoint(x, y))
 
-    def on_select(self): ...
+    def on_select(self):
+        ...
 
-    def on_deselect(self): ...
+    def on_deselect(self):
+        ...
 
     def delete(self):
         for item in self.child_items():
