### 2025-07-30 11a0cd7
fix: do not raise error if .env is not found
