### 2025-08-13 FelipeDefensor/TiLiA#358 @expectopatronum
Thanks for the clarification! I don't really mind about the equal contribution but I am fine if you change it, I'll leave it up to you :)
