### 2025-08-14 FelipeDefensor/TiLiA#362 @expectopatronum
Issues running `pyinstaller tilia.spec`
Hi,

I am currently trying to install the package as outlined in the README (I am using Ubuntu 24.04.3 LTS in WSL, and set up a conda environment) but I get the following error when running `pyinstaller tilia.spec`:

```
ERROR: Aborting build process due to attempt to collect multiple Qt bindings packages: attempting to run hook for 'PyQt6', while hook for 'PyQt5' has already been run! PyInstaller does not support multiple Qt bindings packages in a frozen application - either ensure that the build environment has only one Qt bindings package installed, or exclude the extraneous bindings packages via the module exclusion mechanism (--exclude command-line option, or excludes list in the spec file).
```

I found [this related answer on stackoverflow](https://stackoverflow.com/a/77654484/1117932). They suggest uninstalling Qt6, but since your requirements.txt contain PyQt6, I uninstalled PyQt5.

```
pip uninstall PyQt5
pip uninstall PyQt5_sip
```
But the issue remains. Do you have any suggestions how to proceed? Do you need any additional information (e.g. about my setup)?

Thanks and best regards
Verena

https://github.com/openjournals/joss-reviews/issues/8417
