### 2025-08-14 FelipeDefensor/TiLiA#362 @FelipeDefensor
Hi, Verena.

- Are you using a virtual environment?
- Could you send the output of `pip list`?
