### 2025-08-19 FelipeDefensor/TiLiA#362 @FelipeDefensor
Some of those packages are not related to TiLiA. I'm guessing conda took those from the global Python interpreter, but I can't say for sure as I'm not used to conda. Probably one of those pre-existing packages is either another version of PyQt or has a dependency conflict with it.

You can try creating a clean conda environment with:
```
conda create --no-default-packages -n envname python=3.12
```
That should leave out packages that could cause conflicts. It worked for me.

You can also use the regular `virtualenv` or another package manger.

Uninstalling all packages prefixed with "PyQt" and reinstalling their correct version should also fix it, but I had problems with resolving PyQt version conflicts like that before, so I don't recommend it.
