### 2025-08-19 FelipeDefensor/TiLiA#362 @FelipeDefensor
For reference, here is `pip list` after a clean install:

```
Package             Version     Editable project location
------------------- ----------- ----------------------------------------
anyio               4.10.0
certifi             2025.8.3
chardet             5.2.0
charset-normalizer  3.4.3
colorama            0.4.6
contourpy           1.3.3
cycler              0.12.1
fonttools           4.59.1
h11                 0.16.0
httpcore            1.0.9
httpx               0.28.1
idna                3.10
isodate             0.7.2
joblib              1.5.1
jsonpickle          4.1.1
kiwisolver          1.4.9
lxml                6.0.0
matplotlib          3.10.5
more-itertools      10.7.0
music21             9.7.1
numpy               2.3.2
packaging           25.0
pillow              11.3.0
pip                 25.1
platformdirs        4.3.8
prettytable         3.16.0
pydub               0.25.1
pyparsing           3.2.3
pypdf               4.2.0
PyQt6               6.6.0
PyQt6-Qt6           6.6.0
PyQt6_sip           13.6.0
PyQt6-WebEngine     6.6.0
PyQt6-WebEngine-Qt6 6.6.0
python-dateutil     2.9.0.post0
python-dotenv       1.1.1
requests            2.32.5
sentry-sdk          2.35.0
setuptools          78.1.1
six                 1.17.0
sniffio             1.3.1
TiLiA               0.5.14      C:\Users\username\dev\tmp\tilia-conda-test
urllib3             2.5.0
wcwidth             0.2.13
webcolors           24.11.1
wheel               0.45.1
```
