### 2025-08-20 FelipeDefensor/TiLiA#363 @stefan-balke
Review on manuscript
I will use this thread to collect my thoughts while reading.

Part of: https://github.com/openjournals/joss-reviews/issues/8417
