### 2025-08-20 FelipeDefensor/TiLiA#363 @stefan-balke
ll. 23 please cite 2-3 other works.

May include:
Songle: https://songle.jp/
Sonic Visualiser
Works by Meinard Müller on the SyncPlayer and or related scientifc publications on corpus analyis in music structure

Update: Somewhat resolved with footnote 2 and 3. However, scientific work is missing a bit.
