### 2025-08-20 FelipeDefensor/TiLiA#363 @stefan-balke
l. 48 For the mentioned classes, please provide a class diagram and describe the essential methods and properties
