### 2025-08-20 FelipeDefensor/TiLiA#363 @stefan-balke
ll. 67-74 It would be interesting on how you modeled the timelines and on what abstract level you linked them.

For instance, in our works, e.g. http://mir.audiolabs.uni-erlangen.de/2017-GI-DemoWalkuere/main_view/WWV086B-1# (click on "Schafer", the other videos seem to be unavailable...), we use measures as our base timeline and link everything to that timeline. There is a corresponding publication explaining the concept: https://audiolabs-erlangen.de/content/05_fau/assistant/00_balke/01_publications/2017_BalkeHSAMM_OpernInterface_GI.pdf
Sorry, it is in German, but ChatGPT might resolve that :-)
