### 2025-08-20 FelipeDefensor/TiLiA#363 @stefan-balke
Add a section with an Outlook. What are your goals with the platform. What are possible use cases? Do you want to integrate existing datasets? Do you want to attract musicologists? How could the resulting data be used and share?
