### 2025-08-25 FelipeDefensor/TiLiA#363 @FelipeDefensor
@stefan-balke, I will post the adjusted sections here so you can review them. Once we are happy with them, I will add them to the paper.
