### 2025-08-25 FelipeDefensor/TiLiA#363 @FelipeDefensor
> ll. 67-74 It would be interesting on how you modeled the timelines and on what abstract level you linked them.
>
> For instance, in our works, e.g. http://mir.audiolabs.uni-erlangen.de/2017-GI-DemoWalkuere/main_view/WWV086B-1# (click on "Schafer", the other videos seem to be unavailable...), we use measures as our base timeline and link everything to that timeline. There is a corresponding publication explaining the concept: https://audiolabs-erlangen.de/content/05_fau/assistant/00_balke/01_publications/2017_BalkeHSAMM_OpernInterface_GI.pdf Sorry, it is in German, but ChatGPT might resolve that :-)

Kein Problem. Das ist die perfekte Gelegenheit, mein Deutsch zu üben ;).

At the moment, we use a naive approach: all timeline component are positioned with float timestamps. I hope to migrate to something more robust like what you proposed in the paper in the future.

Here is an explanation for the paper:

> We connect timelines and its components by using the loaded media as a "base layer" to which all timelines refer. Timeline components have either `time` or `start_time` and `end_time` timestamps which positions them in relation to the underlying media. If a different media is loaded, the user has the option to either uniformly scale timelines to the new media or to keep them at their original position (deleting out-of-bounds components if necessary).
If a beat timeline is present, the software can map timestamps to metric positions, also providing a metric-based interface for importing, exporting and interacting with timelines.
