### 2025-08-25 FelipeDefensor/TiLiA#363 @FelipeDefensor
> l. 48 For the mentioned classes, please provide a class diagram and describe the essential methods and properties

Would this work? It is a *greatly* simplified version of what actually happens in the codebase, as we don't have much space.

Acoompanying text for the article:

> Figure X is a *simplified* diagram for the main timeline base classes. `Timeline` acts as container for `TimelineComponent`s. Both act as models of `TimelineUI` and `TimelineUIElement` are views. The `body` property of `TimelineUIElement` is a subclass of `QGraphicsItem`, a Qt base class whose display can be freely customised. The bottom part of the image shows concrete subclasses for the marker timeline type.
