### 2025-08-25 FelipeDefensor/TiLiA#363 @FelipeDefensor
> Add a section with an Outlook. What are your goals with the platform. What are possible use cases? Do you want to integrate existing datasets? Do you want to attract musicologists? How could the resulting data be used and share?

The TiLiA desktop application aims to provide a user-friendly annotation tool both for MIR researchers and musicologists. It can be used to create large annotation datasets or detailed, interactive analyses of individual pieces. It can also be used in the classroom to better connect theoretical concepts with audiovisual material or as common interface for analytical assignments.

In the future we hope to convert existing open-source datasets
and integrate them into the web platform, providing a centralised hub were diversified audiovisual data can be queried and interacted with in a uniform manner.
We also plan to improve audio processing and score displaying capabilities and to add more timeline types (e.g. a 2-dimensional graph timeline) to offer more diversied visualisations.
