### 2025-08-28 FelipeDefensor/TiLiA#365 @expectopatronum
avconv / ffmpeg not mentioned as a requirement
Hi,

when running `python -m tilia.main --user-interface cli` I got the error `/home/myusername/anaconda3/envs/py312-tilia/lib/python3.12/site-packages/pydub/utils.py:170: RuntimeWarning: Couldn't find ffmpeg or avconv - defaulting to ffmpeg, but may not work`. I installed ffmpeg and this seems resolved now. To make the installation instructions complete (as the [review](https://github.com/openjournals/joss-reviews/issues/8417) asks for "Installation: Does installation proceed as outlined in the documentation?" / "Installation instructions: Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.") I would like to see this step mentioned in the installation instructions.

Best regards
Verena
