### 2025-08-28 FelipeDefensor/TiLiA#366 @expectopatronum
ImportError: libsmime3.so: cannot open shared object file / ImportError: libxkbfile.so.1: cannot open shared object file
Hi,

when running `python -m tilia.main --user-interface cli` I receive the following error:

```
Traceback (most recent call last):
  File "<frozen runpy>", line 198, in _run_module_as_main
  File "<frozen runpy>", line 88, in _run_code
  File "/home/myusername/Repos/tilia-desktop/tilia/main.py", line 1, in <module>
    from tilia.boot import boot
  File "/home/myusername/Repos/tilia-desktop/tilia/boot.py", line 9, in <module>
    from tilia.app import App
  File "/home/myusername/Repos/tilia-desktop/tilia/app.py", line 14, in <module>
    from tilia.media.loader import load_media
  File "/home/myusername/Repos/tilia-desktop/tilia/media/loader.py", line 7, in <module>
    from tilia.media.player import Player
  File "/home/myusername/Repos/tilia-desktop/tilia/media/player/__init__.py", line 2, in <module>
    from .youtube import YouTubePlayer
  File "/home/myusername/Repos/tilia-desktop/tilia/media/player/youtube.py", line 13, in <module>
    from PyQt6.QtWebEngineWidgets import QWebEngineView
ImportError: libsmime3.so: cannot open shared object file: No such file or directory
```

After some search I found [this](https://forum.posit.co/t/installation-error-cannot-find-libsmime3-so/30646/3) and installed the mentioned library: `sudo apt install libnss3`

Now I get this error:

```
Traceback (most recent call last):
  File "<frozen runpy>", line 198, in _run_module_as_main
  File "<frozen runpy>", line 88, in _run_code
  File "/home/myusername/Repos/tilia-desktop/tilia/main.py", line 1, in <module>
    from tilia.boot import boot
  File "/home/myusername/Repos/tilia-desktop/tilia/boot.py", line 9, in <module>
    from tilia.app import App
  File "/home/myusername/Repos/tilia-desktop/tilia/app.py", line 14, in <module>
    from tilia.media.loader import load_media
  File "/home/myusername/Repos/tilia-desktop/tilia/media/loader.py", line 7, in <module>
    from tilia.media.player import Player
  File "/home/myusername/Repos/tilia-desktop/tilia/media/player/__init__.py", line 2, in <module>
    from .youtube import YouTubePlayer
  File "/home/myusername/Repos/tilia-desktop/tilia/media/player/youtube.py", line 13, in <module>
    from PyQt6.QtWebEngineWidgets import QWebEngineView
ImportError: libxkbfile.so.1: cannot open shared object file: No such file or directory
```

Are there any other dependencies I am missing?

Thansk a lot & best regards
Verena
(reviewer at [8417](https://github.com/openjournals/joss-reviews/issues/8417))
