### 2025-08-28 FelipeDefensor/TiLiA#367 @expectopatronum
TiLiA visualizations not found
Hi,

I am currently reviewing the documentation of the software for review [8417](https://github.com/openjournals/joss-reviews/issues/8417) (I am assuming the README of the repo is the documentation, please correct me if I am wrong) and I tried to open the linked visualizations ([1](https://tilia-app.com/viewer/1/), [2](https://tilia-app.com/viewer/2/), [3](https://tilia-app.com/viewer/3/)) and I get:

(pasting the screenshot in case it's something temporary).

Thanks and best regards
Verena
(reviewer at [8417](https://github.com/openjournals/joss-reviews/issues/8417))
