### 2025-08-29 FelipeDefensor/TiLiA#365 @FelipeDefensor
Thanks for pointing that out. I got used to ignoring that warning, as it didn't seem to cause any issues, but now that I think about it, I don't think I properly tested TiLiA in a machine without ffmpeg.

Will work on this starting Monday.
