### 2025-09-03 5d2d51a
refactor: fix typing and enumeration
