### 2025-09-03 14874cc
format: auto-format abstract methods
