+++ tilia/media/player/base.py
@@ -295,49 +295,64 @@ class Player(ABC):
             self.load_media(media_path)
 
     @abstractmethod
-    def _engine_pause(self) -> None: ...
+    def _engine_pause(self) -> None:
+        ...
 
     @abstractmethod
-    def _engine_unpause(self) -> None: ...
+    def _engine_unpause(self) -> None:
+        ...
 
     @abstractmethod
-    def _engine_get_current_time(self) -> float: ...
+    def _engine_get_current_time(self) -> float:
+        ...
 
     @abstractmethod
-    def _engine_stop(self): ...
+    def _engine_stop(self):
+        ...
 
     @abstractmethod
-    def _engine_seek(self, time: float) -> None: ...
+    def _engine_seek(self, time: float) -> None:
+        ...
 
     @abstractmethod
-    def _engine_unload_media(self) -> None: ...
+    def _engine_unload_media(self) -> None:
+        ...
 
     @abstractmethod
-    def _engine_load_media(self, media_path: str) -> None: ...
+    def _engine_load_media(self, media_path: str) -> None:
+        ...
 
     @abstractmethod
-    def _engine_play(self) -> None: ...
+    def _engine_play(self) -> None:
+        ...
 
     @abstractmethod
-    def _engine_get_media_duration(self) -> float: ...
+    def _engine_get_media_duration(self) -> float:
+        ...
 
     @abstractmethod
-    def _engine_exit(self) -> float: ...
+    def _engine_exit(self) -> float:
+        ...
 
     @abstractmethod
-    def _engine_set_volume(self, volume: int) -> None: ...
+    def _engine_set_volume(self, volume: int) -> None:
+        ...
 
     @abstractmethod
-    def _engine_set_mute(self, is_muted: bool) -> None: ...
+    def _engine_set_mute(self, is_muted: bool) -> None:
+        ...
 
     @abstractmethod
-    def _engine_try_playback_rate(self, playback_rate: float) -> None: ...
+    def _engine_try_playback_rate(self, playback_rate: float) -> None:
+        ...
 
     @abstractmethod
-    def _engine_set_playback_rate(self, playback_rate: float) -> None: ...
+    def _engine_set_playback_rate(self, playback_rate: float) -> None:
+        ...
 
     @abstractmethod
-    def _engine_loop(self, is_looping: bool) -> None: ...
+    def _engine_loop(self, is_looping: bool) -> None:
+        ...
 
     def __repr__(self):
         return f"{type(self)}-{id(self)}"
