### 2025-09-03 5c132ba
feat: do not require ffmpeg
We are using soundfile instead, which can handle the simple operations we require. I am not sure it supports *all* the formats ffmpeg does.
