+++ setup.cfg
@@ -9,9 +9,9 @@ packages = tilia
 install_requires =
     lxml>=5.3.0
     platformdirs>=4.0.0
-    pydub>=0.25.1
     prettytable>=3.7.0
     music21>=9.1
+    numpy~=1.26.4
     PyQt6==6.6.0
     PyQt6-Qt6==6.6.0
     PyQt6-sip==13.6.0
@@ -20,6 +20,7 @@ install_requires =
     python-dotenv>=1.0.1
     pypdf~=4.2.0
     sentry-sdk>=2.21.0
+    soundfile~=0.13.1
     httpx~=0.28.1
     isodate~=0.7.2
     colorama~=0.4.6
