### 2025-09-03 ff9a59b
build: bump pypdf to 6.0.0
pydpdf<6.0.0 has a known vulnerability.
