### 2025-09-03 FelipeDefensor/TiLiA#366 @FelipeDefensor
Added the dependencies to README.md ([1764ac4](https://github.com/TimeLineAnnotator/desktop/commit/1764ac4cdac2068a40908a6b289ed7aae98b04cd)) and to the website installation instructions (https://tilia-app.com/help/installation/).
