### 2025-09-03 FelipeDefensor/TiLiA#365 @FelipeDefensor
#369 removes ffmpeg from the dependencies. We will probably need it in the future, at which point we can document the dependency properly.
