### 2025-09-05 FelipeDefensor/TiLiA#370 @expectopatronum
Failed to create wl_display (No such file or directory)
When running `python -m tilia.main --user-interface cli` I now get

```
Failed to create wl_display (No such file or directory)
qt.qpa.plugin: Could not load the Qt platform plugin "wayland" in "" even though it was found.
qt.qpa.plugin: Could not load the Qt platform plugin "xcb" in "" even though it was found.
This application failed to start because no Qt platform plugin could be initialized. Reinstalling the application may fix this problem.

Available platform plugins are: wayland, xcb, wayland-egl, vkkhrdisplay, minimalegl, eglfs, vnc, offscreen, linuxfb, minimal.

Aborted (core dumped)
```

I installed the dependency qt6-wayland as suggested in some internet forum, but I still get this error.

Best regards & sorry for the many problems
Verena (Review [8417](https://github.com/openjournals/joss-reviews/issues/8417))
