### 2025-09-05 FelipeDefensor/TiLiA#370 @FelipeDefensor
Oh, sorry for that once more. What distro and version are you running? Did you try running the binary in the website (https://tilia-app.com/downloads) instead of building from source?

Here are some things that might help:
- Setting the `QT_QPA_PLATFORM` environment variable to `wayland`
- Setting the `QT_QPA_PLATFORM` environment variable to `xcb`
