### 2025-09-05 #8417 @FelipeDefensor
Dear @expectopatronum,
The first [issue you raised](https://github.com/TimeLineAnnotator/desktop/issues/38) mentioned the Windows Defender screen. Did you try clicking through it, as suggested in the [installation instructions](https://tilia-app.com/help/installation/)? In case you're not confortable with disregarding the warning, did you try building the app from source on Windows?

Later tickets mention dependency issues in Linux, which I believe we have fixed. Regarding the last ticket, I just sent some suggestions as to how to fix it and am waiting on more information about your setup so I can try to reproduce the problem. I have also suggested trying to run the executable available [in our website](https://tilia-app.com/downloads) instead of building from source while we figure that out.
