### 2025-10-21 #8417 @expectopatronum
> Thanks [@FelipeDefensor](https://github.com/FelipeDefensor). [@expectopatronum](https://github.com/expectopatronum) – did [@FelipeDefensor](https://github.com/FelipeDefensor)'s suggestions help at all here?

Thank your for the suggestions. I had actually tried all those suggestions before I wrote my previous comment, and none of them worked for me. At this point, I am not sure I am the right person to conduct this review, as I have already spent several hours trying to get it to work and I do not think it should be that difficult to install.
