### 2025-10-27 FelipeDefensor/TiLiA#363 @FelipeDefensor
@stefan-balke, when you have the time, please take a look at the proposed changes.
