### 2025-10-27 1337c67
refactor: rename import function
