+++ tests/parsers/csv/test_markers_from_csv.py
@@ -9,7 +9,7 @@ from tests.parsers.csv.common import assert_in_errors
 from tilia.ui import actions
 from tilia.ui.actions import TiliaAction
 from tilia.ui.format import format_media_time
-from tilia.ui.ui_import import on_import_from_csv
+from tilia.ui.ui_import import on_import_to_timeline
 
 
 def patch_import(by: Literal["time", "measure"], tl, data) -> tuple[str, list[str]]:
@@ -18,7 +18,7 @@ def patch_import(by: Literal["time", "measure"], tl, data) -> tuple[str, list[st
 
     def mock_import(timeline_uis, tlkind):
         nonlocal status, errors
-        status, errors = on_import_from_csv(timeline_uis, tlkind)
+        status, errors = on_import_to_timeline(timeline_uis, tlkind)
         return status, errors
 
     with (
@@ -28,7 +28,7 @@ def patch_import(by: Literal["time", "measure"], tl, data) -> tuple[str, list[st
         ),
         patch.object(QFileDialog, "exec", return_value=True),
         patch.object(QFileDialog, "selectedFiles", return_value=[Path()]),
-        patch("tilia.ui.qtui.on_import_from_csv", side_effect=mock_import),
+        patch("tilia.ui.qtui.on_import_to_timeline", side_effect=mock_import),
         patch("builtins.open", mock_open(read_data=data)),
     ):
         actions.trigger(TiliaAction.IMPORT_CSV_MARKER_TIMELINE)
