+++ tilia/ui/qtui.py
@@ -48,7 +48,7 @@ from .menus import (
 )
 from .options_toolbar import OptionsToolbar
 from .player import PlayerToolbar
-from .ui_import import on_import_from_csv
+from .ui_import import on_import_to_timeline
 from .windows.manage_timelines import ManageTimelines
 from .windows.metadata import MediaMetadataWindow
 from .windows.about import About
@@ -186,10 +186,10 @@ class QtUI:
             (Post.REQUEST_CLEAR_UI, self.on_clear_ui),
             (Post.TIMELINE_KIND_INSTANCED, self.on_timeline_kind_change),
             (Post.TIMELINE_KIND_NOT_INSTANCED, self.on_timeline_kind_change),
-            (Post.IMPORT_CSV, self.on_import_from_csv),
+            (Post.IMPORT_CSV, self.on_import_to_timeline),
             (
                 Post.IMPORT_MUSICXML,
-                partial(self.on_import_from_csv, TlKind.SCORE_TIMELINE),
+                partial(self.on_import_to_timeline, TlKind.SCORE_TIMELINE),
             ),
             (Post.DISPLAY_ERROR, display_error),
             (Post.UI_EXIT, self.exit),
@@ -407,9 +407,9 @@ class QtUI:
     def on_website_help_open(self):
         QDesktopServices.openUrl(QUrl(f"{constants.WEBSITE_URL}/help"))
 
-    def on_import_from_csv(self, tl_kind: TlKind):
+    def on_import_to_timeline(self, tl_kind: TlKind):
         prev_state = get(Get.APP_STATE)
-        status, errors = on_import_from_csv(self.timeline_uis, tl_kind)
+        status, errors = on_import_to_timeline(self.timeline_uis, tl_kind)
 
         if status == "failure":
             post(Post.APP_STATE_RESTORE, prev_state)
