+++ tilia/errors.py
@@ -112,7 +112,6 @@ INVALID_MEASURE_FRACTION = Error(
     "Invalid measure fraction",
     "Measure fraction '{}' is not valid. Using closest valid value.",
 )
-INVALID_CSV_ERROR = Error("Import error", "'{}' is not a valid UTF-8 encoded CSV file.")
 SCORE_STAFF_ID_ERROR = Error(
     "Error creating score timeline",
     "Duplicate or missing staff numbers. Staff numbers found: {}.\nDeleting timeline.",
