### 2025-10-27 599d6bb
fix: use utf-8 encoding to import musicxml files
