### 2025-11-04 83bb638
fix: missing .env file is raising extra warnings in Sentry during dev.
.env is already in .gitignore, any local changes made should not be tracked.
Reverts 11a0cd 661ff2.
