### 2025-11-04 7c189f0
ci: test for python 3.13
removal of pydub (5c132b) and related audioop should allow 3.13 to compile.
