+++ .github/workflows/run-tests.yml
@@ -24,8 +24,7 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        # TODO: test for 3.13
-        python-version: ['3.10', '3.11', '3.12']
+        python-version: ['3.10', '3.11', '3.12', '3.13']
         os: [macos-latest, ubuntu-latest, windows-latest]
         include:
         - os: ubuntu-latest
