+++ tilia/ui/windows/inspect.py
@@ -49,6 +49,7 @@ class Inspect(QDockWidget):
 
     def __init__(self, main_window) -> None:
         super().__init__(main_window)
+        self.setObjectName("inspector")
         self.setWindowTitle("Inspector")
         self.setMinimumWidth(250)
         self.setFeatures(
