### 2025-11-13 #8417 @stefan-balke
Hi everybody,

sorry, this has been sitting on my desk for too long and the other comments did not motivate further.

I was able to run it on my mac machine using a virtual environment in `micromamba`, so basially `conda`.
The package `icecream` was missing. Furthermore, the README should state how to run the GUI mode instead of mentioning only the CLI mode.

I am now ready to play around with the functionality. However, I am missing a starter example. @FelipeDefensor the links in the README are broken. Can you point me to some files that would work out of the box?

Thanks
Stefan
