### 2025-11-13 #8417 @FelipeDefensor
Hello @stefan-balke,

Did you build from the `main` branch? If so please use the `dev` branch, as that has the fixes made during this review. The links the README are fixed there.

> README should state how to run the GUI mode instead of mentioning only the CLI mode.

I agree, will add that.
