### 2025-11-15 9d764f5
ci: run tests for newer python versions
