### 2025-11-17 d6a5b84
build: PyQt -> PySide
