+++ tilia/ui/windows/metadata_edit_fields.py
@@ -1,6 +1,6 @@
 import functools
 
-from PyQt6.QtWidgets import QDialog, QVBoxLayout, QTextEdit, QDialogButtonBox
+from PySide6.QtWidgets import QDialog, QVBoxLayout, QTextEdit, QDialogButtonBox
 
 from tilia.requests import Get, get
 
