+++ tilia/ui/dialogs/by_time_or_by_measure.py
@@ -1,5 +1,5 @@
-from PyQt6.QtCore import Qt
-from PyQt6.QtWidgets import (
+from PySide6.QtCore import Qt
+from PySide6.QtWidgets import (
     QDialog,
     QLabel,
     QRadioButton,
