+++ tilia/ui/timelines/slider/timeline.py
@@ -1,9 +1,9 @@
 from __future__ import annotations
 from typing import TYPE_CHECKING
 
-from PyQt6.QtCore import QRectF, QLineF, Qt
-from PyQt6.QtGui import QPen, QColor, QBrush
-from PyQt6.QtWidgets import (
+from PySide6.QtCore import QRectF, QLineF, Qt
+from PySide6.QtGui import QPen, QColor, QBrush
+from PySide6.QtWidgets import (
     QGraphicsEllipseItem,
     QGraphicsLineItem,
     QGraphicsDropShadowEffect,
