### 2025-11-24 9c5883f
ci: test does not require screen definition
already set to offscreen in env.
