+++ .github/workflows/run-tests.yml
@@ -35,7 +35,6 @@ jobs:
           path: ~\AppData\Local\pip\Cache
     timeout-minutes: 30
     env:
-      DISPLAY: ":99.0"
       QT_SELECT: "qt6"
 
     steps:
@@ -51,7 +50,6 @@ jobs:
         sudo apt-get update
         sudo apt-get install -y xvfb libxkbcommon-x11-0 libxcb-icccm4 libxcb-image0 libxcb-keysyms1 libxcb-randr0 libxcb-render-util0 libxcb-xinerama0 libxcb-xinput0 libxcb-xfixes0 libxcb-shape0 libglib2.0-0 libgl1-mesa-dev libpulse-dev
         sudo apt-get install '^libxcb.*-dev' libx11-xcb-dev libglu1-mesa-dev libxrender-dev libxi-dev libxkbcommon-dev libxkbcommon-x11-dev
-        sudo /usr/bin/Xvfb $DISPLAY -screen 0 1280x1024x24 &
 
     - uses: actions/cache@v4
       with:
