### 2025-11-24 0ad3c3e
ci: min ver PySide 6.8 set for long-term support
(see Qt's support page)
