+++ requirements.txt
@@ -3,9 +3,9 @@ prettytable>=3.7.0
 lxml>=5.3.0
 music21>=9.1
 numpy~=1.26.4
-PySide6>=6.9.0
-PySide6_Addons>=6.9.0
-PySide6_Essentials>=6.9.0
+PySide6>=6.8.0
+PySide6_Addons>=6.8.0
+PySide6_Essentials>=6.8.0
 python-dotenv>=1.0.1
 pypdf>=6.0.0
 httpx~=0.28.1
