+++ setup.cfg
@@ -12,9 +12,9 @@ install_requires =
     prettytable>=3.7.0
     music21>=9.1
     numpy~=1.26.4
-    PySide6>=6.10.0
-    PySide6_Addons>=6.10.0
-    PySide6_Essentials>=6.10.0
+    PySide6>=6.8.0
+    PySide6_Addons>=6.8.0
+    PySide6_Essentials>=6.8.0
     python-dotenv>=1.0.1
     pypdf~=4.2.0
     sentry-sdk>=2.21.0
