### 2025-11-24 dcb44af
ci: PySide currently supports Python [3.10 - 3.13]
