+++ .github/workflows/run-tests.yml
@@ -24,7 +24,7 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        python-version: ['3.10', '3.11', '3.12', '3.13', '3.14']
+        python-version: ['3.10', '3.11', '3.12', '3.13']
         os: [macos-latest, ubuntu-latest, windows-latest]
         include:
         - os: ubuntu-latest
