+++ setup.cfg
@@ -22,7 +22,7 @@ install_requires =
     httpx~=0.28.1
     isodate~=0.7.2
     colorama~=0.4.6
-python_requires = >=3.10
+python_requires=>=3.10,<3.14
 zip_safe = no
 
 [options.extras_require]
