### 2025-11-25 777c51b
chore: remove outdated builder
