### 2025-11-25 f853da7
build: centralise all dependencies in pyproject file
