+++ requirements.txt
@@ -1,20 +1,0 @@
-platformdirs>=4.0.0
-prettytable>=3.7.0
-lxml>=5.3.0
-music21>=9.1
-numpy~=1.26.4
-PySide6>=6.8.0
-PySide6_Addons>=6.8.0
-PySide6_Essentials>=6.8.0
-python-dotenv>=1.0.1
-pypdf>=6.0.0
-httpx~=0.28.1
-isodate~=0.7.2
-soundfile~=0.13.1
-colorama~=0.4.6
-
-pytest>=8.0.0
-pytest-env>=1.1.5
-icecream~=2.1.3
-colorama~=0.4.6
-sentry-sdk>=2.21.0
