+++ setup.cfg
@@ -1,42 +1,3 @@
-[metadata]
-name = TiLiA
-version = 0.5.15
-description = A GUI for creating and visualizing annotations over audio and video files.
-author = Felipe Defensor et al.
-
-[options]
-packages = tilia
-install_requires =
-    lxml>=5.3.0
-    platformdirs>=4.0.0
-    prettytable>=3.7.0
-    music21>=9.1
-    numpy~=1.26.4
-    PySide6>=6.8.0
-    PySide6_Addons>=6.8.0
-    PySide6_Essentials>=6.8.0
-    python-dotenv>=1.0.1
-    pypdf~=4.2.0
-    sentry-sdk>=2.21.0
-    soundfile~=0.13.1
-    httpx~=0.28.1
-    isodate~=0.7.2
-    colorama~=0.4.6
-python_requires=>=3.10,<3.14
-zip_safe = no
-
-[options.extras_require]
-testing =
-    pytest>=8.0.0
-    pytest-env>=1.1.5
-    colorama>=0.4.6
-    icecream>=2.1.0
-ci-tests =
-    coverage>=7.0.0
-    coverage-badge>=1.0.0
-    flake8>=7.0.0
-    pytest-cov>=6.0.0
-
 [flake8]
 max-line-length = 88
 per-file-ignores = __init__.py: F401
