+++ pyproject.toml
@@ -1,10 +1,67 @@
 [build-system]
-requires = ["setuptools>=78.1.1", "wheel"]
+requires = [
+    "setuptools>=78.1.1",
+    "wheel",
+]
 build-backend = "setuptools.build_meta"
 
-[tool.pytest.ini_options]
-env = ['QT_QPA_PLATFORM=offscreen', 'ENVIRONMENT=test']
+[project]
+name = "TiLiA"
+version = "0.5.15"
+description = "A GUI for creating and visualizing annotations over audio and video files."
+license = "GPL-3.0"
+license-files = ["LICENSE"]
+readme = "README.md"
+requires-python = ">=3.10,<3.14"
+dependencies = [
+    "colorama~=0.4.6",
+    "httpx~=0.28.1",
+    "isodate~=0.7.2",
+    "lxml>=5.3.0",
+    "music21>=9.1",
+    "numpy~=1.26.4",
+    "platformdirs>=4.0.0",
+    "prettytable>=3.7.0",
+    "python-dotenv>=1.0.1",
+    "pypdf>=6.0.0",
+    "PySide6>=6.8.0",
+    "PySide6_Addons>=6.8.0",
+    "PySide6_Essentials>=6.8.0",
+    "sentry-sdk>=2.21.0",
+    "soundfile~=0.13.1",
+    "TiLiA"
+]
+
+[[project.authors]]
+name = "Felipe Defensor et al."
+email = "tilia@tilia-app.com"
+
+[project.gui-scripts]
+tilia = "tilia.main:main"
+
+[project.urls]
+Homepage = "https://tilia-app.com"
+Repository = "https://github.com/TimeLineAnnotator/desktop"
+
+[project.optional-dependencies]
+ci-tests = [
+    "coverage>=7.0.0",
+    "coverage-badge>=1.0.0",
+    "flake8>=7.0.0",
+    "pytest-cov>=6.0.0",
+]
+dev = ["icecream>=2.1.0"]
+testing = [
+    "pytest>=8.0.0",
+    "pytest-env>=1.1.5",
+]
 
 [tool.coverage.run]
 source = ["tilia"]
 omit = ["tilia/dev/*"]
+
+[tool.pytest.ini_options]
+env = [
+    "ENVIRONMENT=test",
+    "QT_QPA_PLATFORM=offscreen",
+]
