### 2025-12-03 7686618
refactor: rename
mostly to help nuitka find the entry point
...but with the added benefit of a little less typing day-to-day!
