+++ tilia/__main__.py
@@ -1,9 +1,5 @@
 from tilia.boot import boot
 
 
-def main() -> None:
-    boot()
-
-
 if __name__ == "__main__":
-    main()
+    boot()
