### 2025-12-06 64eaf57
docs: some spelling
