### 2025-12-06 2aba357
docs: update deps
