+++ CONTRIBUTING.md
@@ -28,4 +28,11 @@ If you find a bug, [check](https://github.com/TimeLineAnnotator/desktop/issues) 
 # Suggesting features or enhancements
 We are very much looking for ideas of new features for TiLiA. If you would like to suggest one, [check](https://github.com/TimeLineAnnotator/desktop/issues) to see if your feature has already been requested. If not, [open a new one](https://github.com/TimeLineAnnotator/desktop/issues/new) detailing your suggestion.
 
-You need Python 3.11 to build and test TiLiA. You will also need to have ffmpeg installed to use the export and convert audio features.
+You need Python >=3.10 to build and test TiLiA. You will also need to have ffmpeg installed to use the export and convert audio features.
+
+# Developing for TiLiA
+For a better development experience, we recommend the installation of a few more packages:
+```
+pip install --group dev --group testing
+pre-commit install
+```
