### 2025-12-08 15c5b48
chore: completely remove setup.cfg
Apparently the modern way of development is using pyproject instead of setup.cfg...
Ideally, we might want to consider moving away from black + flake to ruff too.
