+++ .pre-commit-config.yaml
@@ -1,20 +1,27 @@
 repos:
--   repo: https://github.com/psf/black
+  - repo: https://github.com/psf/black
     rev: 22.12.0
     hooks:
-    - id: black
+      - id: black
 
--   repo: https://github.com/pre-commit/pre-commit-hooks
+  - repo: https://github.com/pre-commit/pre-commit-hooks
     rev: v4.4.0
     hooks:
       - id: trailing-whitespace
       - id: end-of-file-fixer
       - id: check-added-large-files
-        args: ['--maxkb=1000']
+        args: ["--maxkb=1000"]
       - id: debug-statements
       - id: check-yaml
 
--   repo: https://github.com/pycqa/flake8
+  - repo: https://github.com/pycqa/flake8
     rev: 7.1.0
     hooks:
-    -   id: flake8
+      - id: flake8
+        args:
+          [
+            "--max-line-length=88",
+            "--per-file-ignores=__init__.py:F401",
+            "--extend-ignore=E203,E501",
+            "--exclude=tilia/dev",
+          ]
