+++ setup.cfg
@@ -1,5 +1,0 @@
-[flake8]
-max-line-length = 88
-per-file-ignores = __init__.py: F401
-extend-ignore = E203,E501
-exclude = tilia/dev
