### 2025-12-08 4b57ee5
refactor: use common load env method
