### 2025-12-08 79ab323
refactor: reduce method calls
env is unlikely to change (at least in deployment) after loading
