### 2025-12-08 86949ad
fix: locating metadata
we are not packaging the toml file in the deployed
