### 2025-12-08 2f81222
refactor: read license file in constants
