### 2025-12-08 0b4a9e1
refactor: defaults to dev
see 1bd7d24
