### 2025-12-08 f4838f5
fix: dir in deployment cannot be chdir'ed into
