### 2025-12-08 aa7bda5
ci: update dependency for toml
