### 2025-12-08 6cf9f84
chore: technical difference
in its deployed state, having the extra packages does not extend any of the app's capabilities
