### 2025-12-09 b78bf9a
fix: more secure dotenv
slightly more secure way of fixing 79f6b55.
.tilia.env now stores the basic env vars.
create another .env file and to store dev specific vars privately.
