### 2025-12-09 93f0a81
build: nuitka
to produce build, install and start up a new application:
```
pip install -e .
tilia-desktop
```
