+++ scripts/deploy.py
@@ -1,0 +1,261 @@
+from colorama import Fore
+import dotenv
+from enum import Enum
+from lxml import etree
+from nuitka.distutils.DistutilsCommands import build as n_build
+import os
+from pathlib import Path
+from subprocess import check_call
+import sys
+import tarfile
+import traceback
+
+
+ref_name = ""
+build_os = ""
+buildlib = Path(__file__).parents[1] / "build"
+toml_file = Path(__file__).parents[1] / "pyproject.toml"
+pkg_cfg = "tilia.nuitka-package.config.yml"
+outdir = Path()
+out_filename = ""
+
+if not toml_file.exists():
+    options = {}
+else:
+    if sys.version_info >= (3, 11):
+        from tomllib import load
+    else:
+        from tomli import load
+
+    with open(toml_file, "rb") as f:
+        options = load(f)
+
+
+class P(Enum):
+    CMD = Fore.BLUE
+    ERROR = Fore.RED
+    OK = Fore.GREEN
+
+
+def _print(text: list[str | list[str]], p_type: P | None = None):
+    if not text:
+        return
+    formatted_text = "\n".join([t.__str__() for t in text])
+    if p_type:
+        formatted_text = p_type.value + formatted_text + Fore.RESET
+    sys.stdout.write(formatted_text + "\n")
+
+
+def _handle_inputs():
+    assert len(sys.argv) == 3, "Incorrect number of inputs"
+    global ref_name, build_os, outdir
+    ref_name = sys.argv[1]
+    build_os = "-".join(
+        [
+            x
+            for x in sys.argv[2].split("-")
+            if not x.replace(".", "", 1).isdigit() and x != "latest"
+        ]
+    )
+    outdir = buildlib / build_os
+
+
+def _get_nuitka_toml() -> list[str]:
+    toml_cmds = []
+    for option, value in options.get("tool", {}).get("nuitka", {}).items():
+        toml_cmds.extend(n_build._parseOptionsEntry(option, value))
+    return toml_cmds
+
+
+def _set_out_filename(name: str, version: str):
+    def _clean_version(v: str) -> list[str]:
+        return v.strip("v ").split(".")
+
+    global out_filename
+    if _clean_version(version) == _clean_version(ref_name):
+        out_filename = f"{name}-v{version}-{build_os}"
+    else:
+        out_filename = f"{name}-v{version}-{ref_name}-{build_os}"
+
+
+def _get_exe_cmd() -> list[str]:
+    name = options.get("project", {}).get("name", "TiLiA")
+    version = options.get("project", {}).get("version", "0")
+    _set_out_filename(name, version)
+    icon_path = Path(__file__).parents[1] / "tilia" / "ui" / "img" / "main_icon.ico"
+    exe_args = [
+        sys.executable,
+        "-m",
+        "nuitka",
+        f"--output-dir={outdir}/exe",
+        f"--product-name={name}",
+        f"--file-version={version}",
+        f"--output-filename={out_filename}",
+        # "--onefile-tempdir-spec={CACHE_DIR}/{PRODUCT}/{VERSION}",
+        f"--macos-app-icon={icon_path}",
+        "--macos-app-mode=gui",
+        f"--macos-app-version={version}",
+        "--windows-console-mode=attach",
+        f"--windows-icon-from-ico={icon_path}",
+        f"--linux-icon={icon_path}",
+    ]
+    if "macos" in build_os:
+        exe_args.append("--mode=app")
+    else:
+        exe_args.append("--mode=onefile")
+
+    return exe_args
+
+
+def _get_sdist() -> Path:
+    for f in outdir.iterdir():
+        if "".join(f.suffixes[-2:]) == ".tar.gz":
+            return f
[…]
