+++ tests/conftest.py
@@ -8,6 +8,10 @@ from PySide6.QtCore import QSettings
 from PySide6.QtWidgets import QApplication
 from colorama import Fore, Style
 
+from tilia.utils import load_dotenv
+
+load_dotenv()
+
 import tilia.constants as constants_module
 import tilia.log as logging_module
 import tilia.settings as settings_module
@@ -27,7 +31,6 @@ from tilia.ui.cli.ui import CLI
 from tilia.ui.windows import WindowKind
 from tilia.requests.get import reset as reset_get
 from tilia.requests.post import reset as reset_post
-from tilia.utils import load_dotenv
 
 try:
     # icecream is a replacement for print()
@@ -49,8 +52,6 @@ pytest_plugins = [
     "tests.timelines.pdf.fixtures",
     "tests.timelines.score.fixtures",
 ]
-
-load_dotenv()
 
 
 class TiliaErrors:
