### 2025-12-09 e2bf4d7
build: nuitka
to produce build, install and start up a new application:
```
pip install -e .
tilia-desktop
```
