+++ scripts/deploy.py
@@ -1,0 +1,220 @@
+from colorama import Fore
+import dotenv
+from enum import Enum
+from nuitka.distutils.DistutilCommands import build as n_build
+import os
+from pathlib import Path
+from subprocess import check_call
+from sys import argv, executable, version_info
+import tarfile
+import traceback
+
+
+outdir = Path(__file__).parent.parent / "build"
+toml_file = Path(__file__).parent.parent / "pyproject.toml"
+pkg_cfg = "tilia.nuitka-package.config.yml"
+out_filename = None
+
+if not toml_file.exists():
+    options = {}
+else:
+    if version_info >= (3, 11):
+        from tomllib import load
+    else:
+        from tomli import load
+
+    with open(toml_file, "rb") as f:
+        options = load(f)
+
+
+class P(Enum):
+    CMD = Fore.BLUE
+    ERROR = Fore.RED
+
+
+def _print(text: list[str | list[str]], p_type: P | None = None):
+    if p_type:
+        print(p_type.value + "\t".join([t.__str__() for t in text]) + Fore.RESET)
+    else:
+        print(*text)
+
+
+def _get_out_filename():
+    assert len(argv) == 2, "Incorrect number of inputs"
+    global out_filename
+    out_filename = argv[1]
+
+
+def _get_nuitka_toml() -> list[str]:
+    toml_cmds = []
+    for option, value in options.get("tool", {}).get("nuitka", {}).items():
+        toml_cmds.extend(n_build._parseOptionsEntry(option, value))
+    return toml_cmds
+
+
+def _get_exe_cmd() -> list[str]:
+    name = options.get("project", {}).get("name", "TiLiA")
+    version = options.get("project", {}).get("version", "0")
+    icon_path = Path(__file__).parents[1] / "tilia" / "ui" / "img" / "main_icon.ico"
+    exe_args = [
+        executable,
+        "-m",
+        "nuitka",
+        f"--output-dir={outdir}",
+        "--report=compilation-report.xml",
+        "--assume-yes-for-downloads",
+        f"--product-name={name}",
+        f"--file-version={version}",
+        "--onefile-tempdir-spec=%CACHE_DIR%/%PRODUCT%/%VERSION%",
+        "--mode=onefile",
+        f"--output-filename={out_filename}",
+        f"--product-name={name}",
+        f"--file-version={version}",
+    ]
+    if "mac" in out_filename:
+        exe_args.extend(
+            [
+                f"--macos-app-icon={icon_path}",
+                "--macos-app-mode=gui",
+                f"--macos-app-version={version}",
+            ]
+        )
+    elif "windows" in out_filename:
+        exe_args.extend(
+            [
+                "--windows-console-mode=disable",
+                f"--windows-icon-from-ico={icon_path}",
+            ]
+        )
+    else:
+        exe_args.extend([f"--linux-icon={icon_path}"])
+
+    return exe_args
+
+
+def _get_sdist() -> str:
+    for f in outdir.iterdir():
+        if "".join(f.suffixes[-2:]) == ".tar.gz":
+            return f
+    raise Exception(f"Could not find sdist in {outdir}:", [*outdir.iterdir()])
+
+
+def _get_main_file() -> tuple[bool, str]:
+    main = _create_lib()
+    _update_yml()
+    return main
+
+
+def _create_lib() -> tuple[Path, str, str]:
+    sdist = _get_sdist()
+    base = ".".join(sdist.name.split(".")[:-2])
+    tilia = f"{base}/tilia"
+    lib = outdir / "Lib"
+
+    ext_data = [
+        f"{base}/{e[3:]}"
+        for e in options.get("tool", {})
+        .get("setuptools", {})
+        .get("package-data", {})
+        .get("tilia", [])
[…]
