+++ pyproject.toml
@@ -1,15 +1,15 @@
 [build-system]
 requires = [
-    "setuptools>=78.1.1",
-    "wheel",
+    "Nuitka<2.8",
+     "setuptools",
+     "toml"
 ]
 build-backend = "setuptools.build_meta"
 
 [project]
 name = "TiLiA"
-version = "0.5.15"
+version = "0.5.15.1"
 description = "A GUI for creating and visualizing annotations over audio and video files."
-license = "GPL-3.0"
 license-files = ["LICENSE"]
 readme = "README.md"
 requires-python = ">=3.10,<3.14"
@@ -38,20 +38,30 @@ name = "Felipe Defensor et al."
 email = "tilia@tilia-app.com"
 
 [project.gui-scripts]
-tilia = "tilia.main:main"
+tilia-desktop = "tilia.__main__:boot"
 
 [project.urls]
 Homepage = "https://tilia-app.com"
 Repository = "https://github.com/TimeLineAnnotator/desktop"
 
 [dependency-groups]
+build = [
+    "build"
+]
 ci-tests = [
     "coverage>=7.0.0",
     "coverage-badge>=1.0.0",
     "flake8>=7.0.0",
     "pytest-cov>=6.0.0",
 ]
-dev = ["icecream>=2.1.0"]
+deploy = [
+    "Nuitka<2.8",
+    "toml"
+]
+dev = [
+    "icecream>=2.1.0",
+    "pre-commit",
+]
 testing = [
     "pytest>=8.0.0",
     "pytest-env>=1.1.5",
@@ -61,8 +71,44 @@ testing = [
 source = ["tilia"]
 omit = ["tilia/dev/*"]
 
+[tool.nuitka]
+enable-plugin = ["pyside6"]
+show-anti-bloat-changes = true
+user-package-configuration-file = "tilia.nuitka-package.config.yml"
+noinclude-qt-translations = true
+include-qt-plugins = ["multimedia"]
+nofollow-import-to = ["*.tests", "*.test"]
+assume-yes-for-downloads = true
+# dry-run = true
+# deployment = true
+
 [tool.pytest.ini_options]
 env = [
     "ENVIRONMENT=test",
     "QT_QPA_PLATFORM=offscreen",
+]
+
+[tool.setuptools.packages.find]
+where = ["."]
+exclude = [
+    "build", "build.*",
+    "dist",
+    "docs", "docs.*",
+    "paper", "paper.*",
+    "scripts",
+    "tests", "tests.*",
+    "venv.*", "venv"
+]
+
+[tool.setuptools.package-data]
+tilia = [
+    "../tilia.env",
+    "../tilia.nuitka-package.config.yml", # because we build from sdist
+    "../LICENSE",
+    "ui/img/*",
+    "ui/fonts/*",
+    "media/player/youtube.html",
+    "media/player/youtube.css",
+    "parsers/score/svg_maker.html",
+    "parsers/score/timewise_to_partwise.xsl",
 ]

[…]
