+++ .github/workflows/build.yml
@@ -32,9 +32,8 @@ jobs:
       - name: Build executable
         shell: bash
         run: |
-          $out_filename = "TiLiA-${{github.ref_name}}-${{matrix.os}}"
-          echo "OUT_FILENAME=$out_filename"
-          python scripts/deploy.py ${{env.OUT_FILENAME}}
+          $out_filename="TiLiA-${{github.ref_name}}-${{matrix.os}}
+          python scripts/deploy.py $out_filename
           export QT_DEBUG_PLUGINS = 1
           echo "Checking dependencies"
           timeout 10 build/$out_filename
