+++ .github/workflows/build.yml
@@ -27,7 +27,7 @@ jobs:
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
-          pip install -e . --group build
+          pip install -e . --group build --group deploy
 
       - name: Build executable
         shell: bash
