+++ .pre-commit-config.yaml
@@ -21,7 +21,7 @@ repos:
         args:
           [
             "--max-line-length=88",
-            "--per-file-ignores=__init__.py:F401",
+            "--per-file-ignores=__init__.py:F401,tilia/__main__:E402,tilia/boot:E402",
             "--extend-ignore=E203,E501",
             "--exclude=tilia/dev",
           ]
