+++ tilia/__main__.py
@@ -1,3 +1,8 @@
+import sys
+from pathlib import Path
+
+sys.path[0] = Path(__file__).parents[1].__str__()
+
 from tilia.boot import boot
 
 
