+++ .github/workflows/build.yml
@@ -32,11 +32,11 @@ jobs:
       - name: Build executable
         shell: bash
         run: |
-          out_filename='TiLiA-${{github.ref_name}}-${{matrix.os}}'
-          echo "python scripts/deploy.py $out_filename" | bash
+          eval "$(python scripts/deploy.py ${{github.ref_name}} ${{matrix.os}} "$@")"
           export QT_DEBUG_PLUGINS=1
           echo "Checking dependencies"
-          echo "timeout 10 build/$out_filename" | bash
+          echo "timeout 10 $out_filename" | bash || true
+          echo "ldd $out_filename" | bash
 
       - name: Upload executable
         uses: actions/upload-artifact@v4
