+++ pyproject.toml
@@ -2,7 +2,6 @@
 requires = [
     "Nuitka<2.8",
      "setuptools",
-     "toml"
 ]
 build-backend = "setuptools.build_meta"
 
