+++ pyproject.toml
@@ -47,7 +47,7 @@ build = [
     "build",
     "ImageIO; sys_platform == 'darwin'",
     "Nuitka<2.8",
-    "pyyaml; sys_platform == 'linux'",
+    "pyyaml",
     "wheel"
 ]
 ci-tests = [
