+++ pyproject.toml
@@ -77,7 +77,6 @@ noinclude-qt-translations = true
 include-qt-plugins = ["multimedia"]
 nofollow-import-to = ["*.tests", "*.test"]
 assume-yes-for-downloads = true
-# dry-run = true
 # deployment = true
 
 [tool.pytest.ini_options]
