### 2026-03-09 #8417 @arfon
@FelipeDefensor – I'm sensitive to the fact that both reviewers have struggled here on repeated attempts to get this software up and running. As a next step, please identify a colleague/collaborator who has not installed this software before to attempt to install the software themselves without your assistance/input. Once I have confirmation from you that this has happened, then I think it would be appropriate for the reviewers to take another look.

At this point I'm going to pause the review and await an update from you confirming that you've satisfied yourself that the install procedures have been smoothed out/made more reliable.
