### 2026-03-25 #8417 @FelipeDefensor
@arfon - The software will be used by the students of a music theory discipline in the upcoming weeks, led by one of our colleagues. I will report back with their experience.
