### 2026-09-01 #8417 @stefan-balke
Installation works now, GUI ist starting, but tilia-app.com is not available and everything links to it.
@FelipeDefensor what is the status?
