### 2026-09-01 #8417 @FelipeDefensor
> Installation works now, GUI ist starting, but tilia-app.com is not available and everything links to it. [@FelipeDefensor](https://github.com/FelipeDefensor) what is the status?

Oh, that is new. Site was definetely up until last week. Thank you for the report.

We are investigating. You can use our other URL in the meantime: https://tilia-ad98d.web.app/
